% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Accrued.R
\docType{methods}
\name{accrued}
\alias{accrued}
\alias{accrued,ContractType,timeDate,missing-method}
\alias{accrued,ContractType,character,missing-method}
\alias{accrued,ContractType,timeDate,RiskFactorConnector-method}
\alias{accrued,ContractType,character,RiskFactorConnector-method}
\title{Derive the accrued interest of a contract per specific time(s)}
\usage{
accrued(object, by, model, ...)

\S4method{accrued}{ContractType,timeDate,missing}(object, by, model)

\S4method{accrued}{ContractType,character,missing}(object, by, model)

\S4method{accrued}{ContractType,timeDate,RiskFactorConnector}(object, by, model)

\S4method{accrued}{ContractType,character,RiskFactorConnector}(object, by, model)
}
\arguments{
\item{object}{object The \code{ContractType} or \code{EventSeries}-object for which to derive the value}

\item{by}{time as per which to compute accrued interest (can be a vector)}

\item{model}{(optional) The \code{RiskFactorConnector} conditional to which events are computed}
}
\value{
A \code{numeric} object representing the accrued interest per time
}
\description{
Derive the accrued interest of a contract per specific time(s)
}
\examples{
# define contract
pam = Pam(what=list(
                    ContractID = "001",
                    Currency = "USD",
                    ContractDealDate = "2012-12-31T00",
                    InitialExchangeDate = "2013-01-01T00",
                    MaturityDate = "2017-12-31T00",
                    StatusDate       = "2012-12-31T00",
                    NotionalPrincipal = 1000,
                    NominalInterestRate = 0.04,
                    CycleOfInterestPayment = "1Y-",
                    ContractRole = "RPA",
                    DayCountConvention = "30E/360"
))

# time-vector
by=timeSequence(from=substring(get(pam, "InitialExchangeDate"), 1, 10),
                to=substring(get(pam,"MaturityDate"), 1, 10),
                by="1 months")

# compute accrued interest for time-vector
accrued(pam, by)

}
\seealso{
\link{ContractType}, \link{EventSeries}, \link{RiskFactorConnector}
}
