% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnalysisDate.R
\docType{class}
\name{AD0-class}
\alias{AD0-class}
\title{A Reference Class to represent an Analysis Date}
\description{
Actus Contract Types (CT) generate post-\code{AD0} cash flows. At the same time,
\code{AD0} marks the point in time as per which (market) risk factors have to be
know in order for a linked CT to be able to derive cash flows.
Hence, both, java CTs and java risk factor objects need to know AD0 in java
ZonedDateTime format. An \code{AD0} object provides this reference.
}
\section{Fields}{

\describe{
\item{\code{jref}}{A rJava java object reference}
}}


\seealso{
\code{\link{generateEvents, processEvents, granularResults}}
}
