% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnalysisDate.R
\docType{methods}
\name{AD0}
\alias{AD0}
\alias{AD0,timeDate-method}
\alias{AD0,character-method}
\title{AD0-class constructor}
\usage{
AD0(date, ...)

\S4method{AD0}{timeDate}(date, ...)

\S4method{AD0}{character}(date, ...)
}
\arguments{
\item{date}{The character date of format \code{YYYY-MM-DDTXX} that will be
used as Analysis Date}

\item{...}{}
}
\value{
An object of class AD0 containing the reference to the Java date
}
\description{
Create an instance of \code{AD0} class passing the date in character format. The
format of the character date in order for it to be converted to a Java
ZonedDateTime date has to be \code{YYYY-MM-DDTXX} where XX is either '00' or '24' 
indicating the timestamp (beginning of day/end of day).
}
\seealso{
\code{\link{processEvents}} and \code{\link{processEvents}}
}
