% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EventSeries.R
\docType{methods}
\name{as.data.frame.EventSeries}
\alias{as.data.frame.EventSeries}
\title{## @include 
#' @export
#' @docType methods
#' @rdname print-methods
setMethod("[<-", signature = c("EventSeries", "numeric", "numeric", "ANY"),
          definition = function(x, i, j, y) {
            x$evs[i,j] = y
            x
          }
)
\code{EventSeries}-to-data.frame method}
\usage{
\method{as.data.frame}{EventSeries}(x)
}
\arguments{
\item{x}{An object of class \code{EventSeries}}
}
\value{
An R-data.frame containing a structured 
        representation of the events in the function argument
}
\description{
Convert an \code{EventSeries} to an R-data.frame
}
\details{
This method extracts all events in an EventSeries, sends
each event's EventDate, EventType, EventValue, and state 
variables from the java-object to R (through rJava) and
structures the data in an R-data.frame object.
}
\examples{
pam = Pam()
set(pam, what=list(
                   ContractID = "001",
                   Currency = "CHF",
                   ContractRole = "RPA",
                   StatusDate       = "2012-12-31T00",
                   ContractDealDate = "2012-12-31T00",
                   InitialExchangeDate = "2013-01-01T00",
                   MaturityDate = "2013-03-31T00",
                   NotionalPrincipal = 1000, 
                   NominalInterestRate = 0.01,
                   DayCountConvention = "30E/360"))
ad = AD0("2012-12-31T24")
evs = events(pam, ad)
class(evs)

evs.df = as.data.frame(evs)
class(evs.df)

}
\seealso{
\code{\link{ContractType, generateEvents, processEvents}}
}
