% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnalysisDate.R
\docType{methods}
\name{as.character.AD0}
\alias{as.character.AD0}
\title{Type-cast an AD0-object to character}
\usage{
\method{as.character}{AD0}(object)
}
\arguments{
\item{object}{An object of class AD0 containing the Java reference to be
casted}
}
\value{
An R-character representing the Java date
}
\description{
Type-cast an AD0-object, i.e. the Java reference, back to an R-character
object of format \code{YYYY-MM-DDTXX}.
}
\examples{
ad0 <- AD0("2014-01-01T00")
as.character.AD0(ad0)

}
\seealso{
\code{\link{AD0}}
}
