% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BaseContract.R
\name{BaseContract}
\alias{BaseContract}
\title{\code{BaseContract}}
\usage{
BaseContract(Dates, CashFlows, ...)
}
\arguments{
\item{Dates}{a character vector representing dates.}

\item{CashFlows}{a numeric vector representing cash flow values}
}
\value{
a contract type of class BaseContract.
}
\description{
Constructor method for classes of type BaseContract representing cash flows. 
The class contains the fields 'Dates' and 'CashFlows', representing the 
dates, as well as values of the cash flows.
}
\examples{
bc <- BaseContract(Dates = c("2019-12-31","2020-12-31"),
                   CashFlows = c(-90,100))
                   
}
