% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Portfolio2.R
\name{c,Portfolio2-method}
\alias{c,Portfolio2-method}
\title{This version:
1) Takes the risk factor connector of the first argument.
Question: is it possible to merge different risk factor environments?
2) The events are lost.
It's difficult to collect them without an explicit loop over all
portfolios.}
\usage{
\S4method{c}{Portfolio2}(x, ...)
}
\description{
This version:
1) Takes the risk factor connector of the first argument.
Question: is it possible to merge different risk factor environments?
2) The events are lost.
It's difficult to collect them without an explicit loop over all
portfolios.
}
