% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cashFlows.R
\name{cashFlows}
\alias{cashFlows}
\title{\code{cashFlows}}
\usage{
cashFlows(x, from, to, riskfactors)
}
\arguments{
\item{x}{The contract for which to calculate cashflows}

\item{from}{Start date of cashflow calculation}

\item{to}{End date of cashflow calculation}

\item{riskfactors}{Object of type \code{YieldCurve} or 
\code{RiskfactorConnector}}
}
\value{
timeSeries of generated cashflows
}
\description{
Function that returns the cashflows generated by object \code{x} in form of
a \code{\link{timeSeries}} object.
}
\examples{
my.account <- bankAccount("2013-12-31", balance = 50000)
cashFlows(my.account, from = "2013-01-01", to = "2019-01-01")

}
