% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelStructure.R
\name{clearEvents}
\alias{clearEvents}
\title{Clears previously computed the analytics "analytics" from the tree "node"}
\usage{
clearEvents(node)
}
\description{
Clears previously computed the analytics "analytics" from the tree "node"
}
