% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ContractType.R, R/MaturityContract.R,
%   R/Annuity.R, R/CompoundContract.R, R/ExoticLinearAmortizer.R,
%   R/ForeignExchangeOutright.R, R/Future.R, R/LinearAmortizer.R,
%   R/NegativeAmortizer.R, R/Option.R, R/PrincipalAtMaturity.R, R/Stock.R,
%   R/Swap.R
\docType{class}
\name{ContractType-class}
\alias{ContractType-class}
\alias{MaturityContract-class}
\alias{Annuity-class}
\alias{CompoundContract-class}
\alias{ExoticLinearAmortizer-class}
\alias{ForeignExchangeOutright-class}
\alias{Future-class}
\alias{LinearAmortizer-class}
\alias{NegativeAmortizer-class}
\alias{Option-class}
\alias{PrincipalAtMaturity-class}
\alias{Stock-class}
\alias{Swap-class}
\title{A Reference Class parent to all ACTUS Contract Types}
\description{
This class is only used in an abstract sense in that it is
not intended to create instances of it. It's use is to 
serve as parent to various implementations of ACTUS CTs as 
a means for designing method-inheritage.

Class \link{MaturityContract} is superclass to all maturity
Contract Types in the ACTUS taxonomy (cf. 
\link{www.projectactus.org}).

Class \link{CompoundContract} is superclass to all compound
(aka combined) Contract Types in the ACTUS taxonomy (cf. 
\link{www.projectactus.org}).
}
\section{Fields}{

\describe{
\item{\code{jref}}{A rJava java object reference}
}}


\seealso{
\code{\link{PrincipalAtMaturity, Annuity, Option}}

\code{\link{ContractType, PrincipalAtMaturity, Annuity}}

\code{\link{ContractType, Swap}}
}
