% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ContractModel.R
\docType{class}
\name{ContractModel-class}
\alias{ContractModel-class}
\title{A Reference Class that represents all Contract Models objects}
\description{
This class is only used in a virtual sense in that it is
not intended to create instances of it. It's use is to 
provide the class structure, i.e. field definitions, to 
various implementations of Contract Models for specific 
Contract Types and as a means for designing method-inheritage.
}
\details{
Generally, a Contract Model has one field that contains a 
reference to a java object of the same class in the java 
ACTUS library (e.g. for 
\code{\link{PrincipalAtMaturityModel}} the referenced 
object will be of class 
org.actus.models.PrincipalAtMaturityModel). Note that
the object "carrying" the attribute values, i.e. the
Contract Model object, lives in Java, or the JVM 
respectively, and not in R. Attribute values may be 
accessed/changed using methods 
\code{\link{get}}/\code{\link{set}} which uses the JNI
interface through R-package rJava.

For a list of available contract attributes use 
method \code{\link{terms}}. For an explanation of the 
attributes refer to the official ACTUS data dictionary
under \link{www.projectactus.org}.
}
\section{Fields}{

\describe{
\item{\code{jref}}{A rJava java object reference}
}}


\seealso{
\code{\link{ContractType, PrincipalAtMaturityModel, 
               AnnuityModel, StockModel}}
}
