% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ContractModel.R
\docType{methods}
\name{CTM}
\alias{CTM}
\alias{CTM,jobjRef-method}
\alias{CTM,character-method}
\alias{CTM,}
\alias{character-method}
\title{\code{ContractModel}-class constructor}
\usage{
CTM(model)

\S4method{CTM}{character}(model)
}
\arguments{
\item{object}{If an object of class \code{jobjRef}, then 
this parameter is expected to be a reference to a Java
contract model and a new R-object of the same class is 
created where the Java reference attached as the 
classes' \code{jref}-field. If a character, then 
\code{object} is expected to be the R-class name of the 
CT for which a contract model is to be instantiated.}
}
\value{
An object of a class extending \code{ContractModel}
}
\description{
Create an instance of an implementation of class 
\code{ContractModel} (e.g. 
\code{\link{PrincipalAtMaturityModel}}, 
\code{\link{StockModel}}, etc).
}
\details{
This constructor is in fact a short cut to the constructors
of the implemented classes such as \code{\link{PamModel}} for 
\code{\link{PrincipalAtMaturityModel}} or 
\code{\link{StkModel}} for \code{\link{StockModel}}. Note 
that it is not possible to instantiate class 
\code{ContractModel} itself but only the implementing 
classes extending \code{ContractModel}.
}
\examples{
# example 1: create a new 'PAM-Model' object
# This example does not work
# pam.model = CTM("PrincipalAtMaturity")

# example 2: attach the reference to a Java 'PAM-Model' object 
#            to a new R-'PAM' object. Note, the new object will
#            refer to the same Java contract.
# This example does not work
# pam.model = CTM("PrincipalAtMaturity")
# set(pam.model, what=list(
#                    ContractID = "001",
#                    Currency = "CHF",
#                    ContractRole = "RPA",
#                    StatusDate       = "2012-12-31T00",
#                    ContractDealDate = "2012-12-31T00",
#                    InitialExchangeDate = "2013-01-01T00",
#                    MaturityDate = "2013-03-31T00",
#                    NotionalPrincipal = 1000, 
#                    NominalInterestRate = 0.01,
#                    DayCountConvention = "30E/360"))
# same.pam.model = CTM(pam.model$jref)

# example 3: create a new 'PAM-Model' object using the
#            CT-specific model constructor method
pam.model = PamModel()

}
\seealso{
\code{\link{ContractModel,PrincipalAtMaturityModel}}
}
