% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiscountingEngine.R, R/CapmEngine.R,
%   R/MultiCurrencyDiscountingEngine.R
\docType{class}
\name{DiscountingEngine-class}
\alias{DiscountingEngine-class}
\alias{CAPMEngine-class}
\alias{MultiCurrencyDiscountingEngine-class}
\title{A \code{\link{ValuationEngine}} implementing discounting of
future cash flows}
\description{
This class represents a \code{\link{ValuationEngine}} that
uses the future cash flow discounting method to compute a
value of a CT. The engine is implemented in Java so that
this class points to an instance of the respective Java
class.

This class represents a \code{link{ValuationEngine}} that
uses Capital Asset Pricing Model (CAPM) method to compute
a value of a CT. The engine is implemented in Java so that
this class points to an instance of the respective Java
class.

This class represents a \code{\link{ValuationEngine}} that
uses the future cash flow discounting method to compute a
value of two legs in different currencies and converts 
these in the desired (target) currency of a CT. Thereby, 
the cash flows to be valued may be in two different 
currencies each being modelled with a separate discounting
curve and discounting spread. Finally, based on a reference 
FX-rate index the values of the two currency-legs are 
converted into the specified target currency. The engine 
is implemented in Java so that this class points to an 
instance of the respective Java class.
}
\section{Fields}{

\describe{
\item{\code{jref}}{A rJava java object reference}

\item{\code{jref}}{A rJava java object reference}

\item{\code{jref}}{A rJava java object reference}
}}


\seealso{
\code{\link{DiscountingEngineModel, ValuationEngine, 
               ValuationEngineModel}}

\code{\link{DiscountingEngine, ValuationEngine, 
               ValuationEngineModel}}

\code{\link{DiscountingEngine, ValuationEngine, 
               ValuationEngineModel}}
}
