% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DynamicYieldCurve.R
\docType{methods}
\name{discountFactors}
\alias{discountFactors}
\alias{discountFactors,YieldCurve,character,missing-method}
\alias{discountFactors,}
\alias{YieldCurve,}
\alias{character,}
\alias{character-method}
\alias{discountFactors,YieldCurve,character,character-method}
\alias{discountFactors,DynamicYieldCurve,character,missing-method}
\alias{DynamicYieldCurve,}
\alias{discountFactors,numeric,character,character-method}
\alias{discountFactors,DynamicYieldCurve,character,character-method}
\alias{missing-method}
\title{Generic method to retrieve discount factors for specific
tenor(s) from a \code{\link{DynamicYieldCurve}} object}
\usage{
discountFactors(object, termEnd, termStart, ...)

\S4method{discountFactors}{YieldCurve,character,missing}(
  object,
  termEnd,
  termStart,
  method = "continuous",
  period = "Y",
  ...
)

\S4method{discountFactors}{YieldCurve,character,character}(
  object,
  termEnd,
  termStart,
  method = "continuous",
  period = "Y",
  ...
)

\S4method{discountFactors}{DynamicYieldCurve,character,missing}(
  object,
  termEnd,
  termStart,
  method = "continuous",
  period = "Y",
  ...
)

\S4method{discountFactors}{numeric,character,character}(
  object,
  termEnd,
  termStart,
  method = "continuous",
  period = "Y",
  ...
)

\S4method{discountFactors}{DynamicYieldCurve,character,character}(
  object,
  termEnd,
  termStart,
  method = "continuous",
  period = "Y",
  ...
)
}
\arguments{
\item{object}{An object of class \code{DynamicYieldCurve} from 
which to extract the discount factors}

\item{termEnd}{The discounting period end date (t2) 
Can be a single value or a vector of tenors.}

\item{termStart}{(optional) The discounting period start
date (t1). Is a single value. If combined 
with a vector for termEnd, then termStart remains
the same for all t2 defined in termEnd.}

\item{...}{Additional parameters:
\itemize{
 \item{"isDateEnd"}{logical indicating whether termEnd is 
       of date (TRUE) or term (FALSE) format. Date format is
       'YYYY-MM-DDTXX' with 'XX'=00 for beginning of day, or
       24 for end of day, and term format is 'IP' with 'I' 
       an integer and 'P' indicating the period (D=days, 
       W=weeks, M=months, Q=quarters, H=halfyears, Y=years).
       Default is isDateEnd=FALSE.}
}}
}
\value{
numeric The discount factor for the defined period
}
\description{
Discount factors for \code{t2}-tenors are extracted from a 
\code{\link{DynamicYieldCurve}} object according to 
\code{df(t0,t1,t2)=exp(-yf(t1,t2)*yield(t1,t2))} where
\itemize{
 \item{"t0"}{is the 'ReferenceDate' of the yield curve}
 \item{"t1"}{marks the discounting period start. If t1>t0, 
             we are in a forward-mode (default: t0=t1, i.e. 
             non-forward)}
 \item{"t2"}{marks the discounting period end}
 \item{"yf(t1,t2)"}{indicates the the year fraction using 
                   Actual/Actual between times t1 and t2}
 \item{"yield(t1,t2)"}{refers to the yield from t1 to t2 
                   implied by the yield curve (forward 
                   rate/yield if t1>t0)}
}
}
\examples{
yc <- DynamicYieldCurve()
tenors <- c("1W", "1M", "6M", "1Y", "2Y", "5Y")
rates <- c(0.001, 0.0015, 0.002, 0.01, 0.02, 0.03)
set(yc, what = list(MarketObjectCode = "YC_Prim",
  Nodes = list(ReferenceDate = "2015-01-01T00", Tenors = tenors, Rates = rates)))

discountFactors(yc, "1Y")
discountFactors(yc, "2016-01-01T00", isDateEnd=TRUE)
discountFactors(yc, "1Y", "2015-07-01T00")

}
\seealso{
\code{\link{rates}}
}
