% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EventSeries.R
\docType{class}
\name{EventSeries-class}
\alias{EventSeries-class}
\title{A Reference Class representing a (Time-) Series of 
Contract Events}
\description{
This class represents a structure collection of Contract
Events. Events are ordered on the timeline according to 
their Event Date and Sequence (cf. ACTUS documentation).
}
\section{Fields}{

\describe{
\item{\code{jref}}{A rJava java object reference}
}}


\seealso{
\code{\link{generateEvents, processEvents}}
}
