% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EventSeries.R, R/Operations.R
\docType{methods}
\name{EventSeries}
\alias{EventSeries}
\alias{EventSeries,missing-method}
\alias{EventSeries,ContractType-method}
\alias{EventSeries,ContractType,timeDate-method}
\alias{EventSeries,}
\alias{missing-method}
\alias{EventSeries,ContractType,AD0-method}
\alias{EventSeries,ContractType,character-method}
\alias{EventSeries,Portfolio,AD0-method}
\alias{EventSeries,EventSeries,missing-method}
\alias{EventSeries,Operations,character-method}
\title{\code{EventSeries}-class constructor}
\usage{
EventSeries(object, ad, ...)

\S4method{EventSeries}{ContractType,timeDate}(object, ad, ...)

\S4method{EventSeries}{ContractType,AD0}(object, ad, ...)

\S4method{EventSeries}{ContractType,character}(object, ad, ...)

\S4method{EventSeries}{Portfolio,AD0}(object, ad, ...)

\S4method{EventSeries}{EventSeries,missing}(object, ad, ...)

\S4method{EventSeries}{Operations,character}(object, ad, ...)
}
\arguments{
\item{object}{(optional) A \code{ContractType}-object from which to extract the EventSeries}
}
\value{
An object of a class \code{EventSeries} containing a
        reference to a java EventSeries object
}
\description{
Create an instance of class \code{EventSeries}
}
\details{
Optionally, a \code{ContractType}-object can be passed on to 
the constructor in which case the java 
reference to object's 'processed event series' is
attached as the classes' \code{jref}-field. 

If no argument is provided, 
then a new empty java event series is created and 
referenced to by the classe's \code{jref} field.
}
\examples{
# example 1: create a new, empty object
# evs = EventSeries() # this constructor doesn't exist

# example 2: create a new object and attach a contract's Event
pam = Pam()
set(pam, what=list(
                   ContractID = "001",
                   Currency = "CHF",
                   ContractRole = "RPA",
                   StatusDate       = "2012-12-31T00",
                   ContractDealDate = "2012-12-31T00",
                   InitialExchangeDate = "2013-01-01T00",
                   MaturityDate = "2013-03-31T00",
                   NotionalPrincipal = 1000, 
                   NominalInterestRate = 0.01,
                   DayCountConvention = "30E/360"))
ad = "2012-12-31T00"
evs = EventSeries(pam, ad)

}
\seealso{
\code{\link{ContractType, generateEvents, processEvents}}
}
