% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelStructure.R
\name{fAnalytics}
\alias{fAnalytics}
\title{general function for computing analytics on a data.tree structure of class Node}
\usage{
fAnalytics(node, ...)
}
\description{
This function computes analytics individually for the leafs of a tree
The analytics to be computed must be passed as first argument.
This function thus subsumes the function of all three specialized 
functions above (which are commented out)
}
