% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ForeignExchangeRate.R
\docType{methods}
\name{FxRate}
\alias{FxRate}
\alias{FxRate-method}
\alias{FxRate,ANY-method}
\title{\code{ForeignExchangeRate}-class constructor}
\usage{
FxRate(...)

\S4method{FxRate}{ANY}(...)
}
\arguments{
\item{...}{}
}
\value{
An object of class \code{ForeignExchangeRate}
}
\description{
Create an instance of \code{ForeignExchangeRate} class.
}
\seealso{
\code{\link{YieldCurve, Index}}
}
