% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YieldCurve2.R
\name{getRateAt2}
\alias{getRateAt2}
\title{Computes the forward rate from time t1 to time t2.}
\usage{
getRateAt2(object, from, to, ...)
}
\description{
Q bwlf:
Is this a helper method or do we need a help text?
A auth:
This function still needs documentation, yes.
}
