% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DynamicYieldCurve.R
\name{getRateAt_old}
\alias{getRateAt_old}
\title{Computes the forward rate from time t1 to time t2.}
\usage{
getRateAt_old(object, from, to, ...)
}
\description{
This function still needs documentation, yes.
}
