% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/immunization.R
\name{immunize}
\alias{immunize}
\title{\code{immunize}}
\usage{
immunize(x, target, yield, isPercentage=TRUE, period, ...)
}
\arguments{
\item{x}{the portfolio of contracts that should be adjusted.}

\item{target}{the liabilities (contract or portfolio) whose value and duration
should be matched.}

\item{yield}{a numeric, indicating the yield used to calculate the duration.}

\item{isPercentage}{a logical, indicating if the 'yield' is expressed as percentage 
(TRUE) or as fraction (FALSE).}

\item{period}{argument currently not used.}

\item{...}{additional parameters to be passed.}
}
\value{
the immunization of the contract or portfolio with respect to the target.
}
\description{
Immunization minimizes the interest rate risk of a fixed-income portfolio by adjusting the 
portfolio duration such that it matches the duration of the liabilities that
should be covered with this portfolio.
}
\examples{
bnd1 <- bond(start="2015-01-01", maturity="30 years", nominal=1000, 
          coupon=0.06, couponFreq="1 year", role="long", variable=FALSE)
bnd2 <- bond(start="2015-01-01", maturity="10 years", nominal=1000, 
          coupon=0.11, couponFreq="1 year", role="long", variable=FALSE)
bnd3 <- bond(start="2015-01-01", maturity="20 years", nominal=1000, 
          coupon=0.09, couponFreq="1 year", role="long", variable=FALSE)
          
ptf <- Portfolio(b1=bnd1,b2=bnd2,b3=bnd3)

target <- bond(start="2015-01-01", maturity="10 years", nominal=1000000, coupon=0, 
               couponFreq="10 year", role="long", variable=FALSE)
immunize(ptf, target, yield=9, from=NULL)

}
