% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YieldCurve.R, R/DynamicYieldCurve.R,
%   R/ReferenceIndex.R
\docType{class}
\name{YieldCurve-class}
\alias{YieldCurve-class}
\alias{DynamicYieldCurve-class}
\alias{ReferenceIndex-class}
\title{A Reference Class extending \code{\link{RiskFactor}} class
and representing a yield curve risk factor}
\description{
Yield curves define a class of market risk factors 
that in the ACTUS model may directly affect future cash 
flows arising from a financial instrument, e.g. Floating-
rate bonds, or are used for valuation of instruments, 
e.g. in discounting.

Yield curves define a class of market risk factors 
that in the ACTUS model may directly affect future cash 
flows arising from a financial instrument, e.g. Floating-
rate bonds, or are used for valuation of instruments, 
e.g. in discounting.

Reference indices define a class of market risk factors 
that in the ACTUS model may directly affect future cash 
flows arising from a financial instrument, e.g. Inflation-
linked bonds, or are used for valuation of instruments, 
e.g. a Stock market index when using CAPM.
}
\section{Fields}{

\describe{
\item{\code{jref}}{A rJava java object reference}

\item{\code{jref}}{A rJava java object reference}

\item{\code{MarketObjectCode}}{character name of the risk factor}

\item{\code{TimeSeries}}{data.frame representating time series data}
}}


\examples{
yc <- YieldCurve()
tenors <- c("1W", "1M", "6M", "1Y", "2Y", "5Y")
rates <- c(0.001, 0.0015, 0.002, 0.01, 0.02, 0.03)
set(yc, what = list(MarketObjectCode = "YC_Prim",
  Nodes = list(ReferenceDate = "2015-01-01T00", Tenors = tenors, Rates = rates)))

get(ind, "MarketObjectCode")
get(yc, "ReferenceDate")
get(yc, "Tenors")
get(yc, "Rates")

yc <- DynamicYieldCurve()
tenors <- c("1W", "1M", "6M", "1Y", "2Y", "5Y")
rates <- c(0.001, 0.0015, 0.002, 0.01, 0.02, 0.03)
set(yc, what = list(MarketObjectCode = "YC_Prim",
  Nodes = list(ReferenceDate = "2015-01-01T00", Tenors = tenors, Rates = rates)))

get(ind, "MarketObjectCode")
get(yc, "ReferenceDate")
get(yc, "Tenors")
get(yc, "Rates")


# create an ReferenceIndex object
ind <- Index()

# define time stamps and values
times <- c("2015-01-01", "2016-01-01", "2017-01-01", "2018-01-01",
           "2019-01-01")
values <- c(100, 110, 120, 130, 140)

# set the MarketObjectCode and Data
set(ind, what = list(MarketObjectCode = "CHF_SMI",
                     Data = list(Dates = times, 
                                       Values = values)))

# get MarketObjectCode
get(ind, "MarketObjectCode")

# get values of the risk factor at certain times
valueAt(ind, "2016-01-01")
valueAt(ind, c("2016-01-01", "2018-07-01", "2018-07-01"))

}
\seealso{
\code{\link{RiskFactor, ReferenceIndex, ForeignExchangeRate}}

\code{\link{RiskFactor, ReferenceIndex, ForeignExchangeRate}}

\code{\link{RiskFactor, YieldCurve, ForeignExchangeRate}}
}
