% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReferenceIndex.R
\docType{methods}
\name{Index}
\alias{Index}
\alias{Index-method}
\alias{Index,numeric,ANY,character-method}
\title{\code{ReferenceIndex}-class constructor}
\usage{
Index(data, charvec, label, ...)

\S4method{Index}{numeric,ANY,character}(data, charvec, label, ...)
}
\arguments{
\item{...}{}
}
\value{
An object of class \code{ReferenceIndex} 
         containing the reference to the Java object
}
\description{
Create an instance of \code{ReferenceIndex} class. The 
constructor will also create an instance of the respective
Java class in the running JVM.
}
\seealso{
\code{\link{YieldCurve, ForeignExchangeRate}}
}
