% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Institution.R
\name{Institution}
\alias{Institution}
\alias{institution}
\title{Class that contains the whole model of an enterprise or institution.}
\usage{
institution(name, cashcollect = TRUE, ...)
}
\description{
The class inherits from class \code{Node} in package \code{data.tree}
It contains a hierarchical model structure.
The upper level is predefined with the nodes "Active", "Passive" and
"Operations".
}
\section{Super class}{
\code{\link[data.tree:Node]{data.tree::Node}} -> \code{Institution}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-clone}{\code{Institution$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="data.tree" data-topic="Node" data-id="AddChild">}\href{../../data.tree/html/Node.html#method-AddChild}{\code{data.tree::Node$AddChild()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="data.tree" data-topic="Node" data-id="AddChildNode">}\href{../../data.tree/html/Node.html#method-AddChildNode}{\code{data.tree::Node$AddChildNode()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="data.tree" data-topic="Node" data-id="AddSibling">}\href{../../data.tree/html/Node.html#method-AddSibling}{\code{data.tree::Node$AddSibling()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="data.tree" data-topic="Node" data-id="AddSiblingNode">}\href{../../data.tree/html/Node.html#method-AddSiblingNode}{\code{data.tree::Node$AddSiblingNode()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="data.tree" data-topic="Node" data-id="Climb">}\href{../../data.tree/html/Node.html#method-Climb}{\code{data.tree::Node$Climb()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="data.tree" data-topic="Node" data-id="Do">}\href{../../data.tree/html/Node.html#method-Do}{\code{data.tree::Node$Do()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="data.tree" data-topic="Node" data-id="Get">}\href{../../data.tree/html/Node.html#method-Get}{\code{data.tree::Node$Get()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="data.tree" data-topic="Node" data-id="Navigate">}\href{../../data.tree/html/Node.html#method-Navigate}{\code{data.tree::Node$Navigate()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="data.tree" data-topic="Node" data-id="Prune">}\href{../../data.tree/html/Node.html#method-Prune}{\code{data.tree::Node$Prune()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="data.tree" data-topic="Node" data-id="RemoveAttribute">}\href{../../data.tree/html/Node.html#method-RemoveAttribute}{\code{data.tree::Node$RemoveAttribute()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="data.tree" data-topic="Node" data-id="RemoveChild">}\href{../../data.tree/html/Node.html#method-RemoveChild}{\code{data.tree::Node$RemoveChild()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="data.tree" data-topic="Node" data-id="Revert">}\href{../../data.tree/html/Node.html#method-Revert}{\code{data.tree::Node$Revert()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="data.tree" data-topic="Node" data-id="Set">}\href{../../data.tree/html/Node.html#method-Set}{\code{data.tree::Node$Set()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="data.tree" data-topic="Node" data-id="Sort">}\href{../../data.tree/html/Node.html#method-Sort}{\code{data.tree::Node$Sort()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="data.tree" data-topic="Node" data-id="initialize">}\href{../../data.tree/html/Node.html#method-initialize}{\code{data.tree::Node$initialize()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Institution$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
