% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YieldCurve.R, R/DynamicYieldCurve.R
\docType{class}
\name{Interpolator-class}
\alias{Interpolator-class}
\alias{Interpolator}
\title{A Reference Class}
\description{
It should be generalized by unsing splinefun. 
In this case, defining a class has real value.
}

