% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Operations.R
\name{Investments}
\alias{Investments}
\title{Investments Contract class definition}
\usage{
Investments(ContractID, pattern, args, ...)
}
\arguments{
\item{pattern}{A function evaluating the pattern of generated investments. 
               The function must implement two arguments:
\itemize{
\item{model} {The \link{RiskFactorConnector} object used when evaluating the pattern}
\item{args} {The \link{args} object used when evaluating the pattern}
}}

\item{args}{The arguments used when evaluating the pattern}
}
\description{
An Investments contract represents any operational activity 
in monetary units within an organization related to investments.
}
\examples{
times = timeSequence(from="2014-01-01", by="3 months", length.out=9)
write.off <- function(times) {
   timeSeries(seq(1000000, 0, length.out=9), times)
   }
invest <- Investments(
              ContractID = "Ops002", Currency = "CHF", 
              pattern = write.off, 
              args = list(times = times))

}
