% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Liquidity.R, R/CurrentAccount.R,
%   R/ModelStructure.R, R/Operations.R
\docType{methods}
\name{liquidity}
\alias{liquidity}
\alias{liquidity,ContractType,timeDate,missing-method}
\alias{liquidity,ContractType,timeBuckets,missing-method}
\alias{liquidity,ContractType,timeDate,character-method}
\alias{liquidity,ContractType,timeBuckets,character-method}
\alias{liquidity,Portfolio,timeDate,missing-method}
\alias{liquidity,Portfolio,timeBuckets,missing-method}
\alias{liquidity,Portfolio,timeBuckets,character-method}
\alias{liquidity,Portfolio,timeDate,character-method}
\alias{liquidity,EventSeries,timeDate,missing-method}
\alias{liquidity,EventSeries,timeBuckets,missing-method}
\alias{liquidity,EventSeries,timeDate,character-method}
\alias{liquidity,EventSeries,timeBuckets,character-method}
\alias{liquidity,eventList,timeDate,missing-method}
\alias{liquidity,eventList,timeBuckets,missing-method}
\alias{liquidity,eventList,timeDate,character-method}
\alias{liquidity,eventList,timeBuckets,character-method}
\alias{liquidity,data.frame,timeDate,character-method}
\alias{liquidity,CurrentAccount,timeDate,missing-method}
\alias{liquidity,CurrentAccount,timeBuckets,missing-method}
\alias{liquidity,CurrentAccount,timeBuckets,character-method}
\alias{liquidity,CurrentAccount,timeDate,character-method}
\alias{liquidity,Node,timeBuckets,character-method}
\alias{liquidity,Operations,timeDate,missing-method}
\alias{liquidity,Operations,timeBuckets,missing-method}
\alias{liquidity,Operations,timeDate,character-method}
\alias{liquidity,Operations,timeBuckets,character-method}
\alias{income,Operations,timeDate,character,ANY,ANY-method}
\title{Derive the liquidity-vector for \code{ContractType}}
\usage{
liquidity(object, by, type, ...)

\S4method{liquidity}{ContractType,timeDate,missing}(object, by, type, digits = 2)

\S4method{liquidity}{ContractType,timeBuckets,missing}(object, by, type, digits = 2)

\S4method{liquidity}{ContractType,timeDate,character}(object, by, type, digits = 2)

\S4method{liquidity}{ContractType,timeBuckets,character}(object, by, type, digits = 2)

\S4method{liquidity}{Portfolio,timeDate,missing}(object, by, type, ...)

\S4method{liquidity}{Portfolio,timeBuckets,missing}(object, by, type, ...)

\S4method{liquidity}{Portfolio,timeBuckets,character}(object, by, type, ...)

\S4method{liquidity}{Portfolio,timeDate,character}(object, by, type, digits = 2, ...)

\S4method{liquidity}{EventSeries,timeDate,missing}(object, by, type, digits = 2)

\S4method{liquidity}{EventSeries,timeBuckets,missing}(object, by, type, digits = 2)

\S4method{liquidity}{EventSeries,timeDate,character}(object, by, type, digits = 2)

\S4method{liquidity}{EventSeries,timeBuckets,character}(object, by, type, digits = 2)

\S4method{liquidity}{eventList,timeDate,missing}(object, by, type, ...)

\S4method{liquidity}{eventList,timeBuckets,missing}(object, by, type, ...)

\S4method{liquidity}{eventList,timeDate,character}(object, by, type, ...)

\S4method{liquidity}{eventList,timeBuckets,character}(object, by, type, ...)

\S4method{liquidity}{data.frame,timeDate,character}(object, by, type, ...)

\S4method{liquidity}{CurrentAccount,timeDate,missing}(object, by, type, digits = 2)

\S4method{liquidity}{CurrentAccount,timeBuckets,missing}(object, by, type, digits = 2)

\S4method{liquidity}{CurrentAccount,timeBuckets,character}(object, by, type, digits = 2)

\S4method{liquidity}{CurrentAccount,timeDate,character}(object, by, type, digits = 2)

\S4method{liquidity}{Node,timeBuckets,character}(object, by, type)

\S4method{liquidity}{Operations,timeDate,missing}(object, by, type, ...)

\S4method{liquidity}{Operations,timeBuckets,missing}(object, by, type, ...)

\S4method{liquidity}{Operations,timeDate,character}(object, by, type, ...)

\S4method{liquidity}{Operations,timeBuckets,character}(object, by, type, ...)

\S4method{income}{Operations,timeDate,character,ANY,ANY}(object, by, type, revaluation.gains, method, ...)
}
\arguments{
\item{object}{The \code{ContractType} or \code{EventSeries}-object for which to derive the liquidity-vector}

\item{by}{A sequence of 'timeDate's providing the target time-axis for the liquidity-vector}

\item{type}{A character representing the type of liquidity (either 'marginal' or 'cumulative')}

\item{...}{Currently unused}
}
\value{
A \code{numeric} object representing the liquidity-vector on the target time-axis
}
\description{
Different liquidity-concepts can be derived for a financial instrument
or the resulting EventSeries, respectively. Currentently, these are
Marginal liquidity and Cumulative liquidity.
}
\details{
Marginal liquidity-vector represents the aggregate cash flows within a
set of user-defined time-intervals. The time-intervals are defined as
a sequence of timeDate-dates. Thereby, the marginal liquidity-vector
gives the net cash flows within the specified time-intervals.

Cumulative liquidity-vector is the cumulative sum over time (-intervals)
of the marginal liquidity-vector.
}
\examples{
pam <- Pam()
set(pam, what=list(
                 ContractID = "001",
                 Currency = "CHF",
                 Calendar = "Weekday",
                 ContractRole = "RPA",               
                 StatusDate       = "2016-05-30T00",
                 ContractDealDate = "2016-05-30T00",
                 InitialExchangeDate = "2016-05-30T00",
                 MaturityDate = "2020-06-01T00",
                 NotionalPrincipal = 1000,
                 NominalInterestRate = 0.05,
                 CycleOfInterestPayment = "1Y-", 
                 PremiumDiscountAtIED = 0.0,
                 DayCountConvention = "30E/360",
                 BusinessDayConvention = "SCF"))
ad <- "2016-06-01T00"

# generate event series
evs = events(pam, ad)

# define target liquidity time axis
by=timeSequence(substring(ad, 1, 10), "2020-06-01", by="1 year")

# derive marginal liquidity for defined time axis
liquidity(evs, by, "marginal")

# derive cumulative liquidity for defined time axis
liquidity(evs, by, "cumulative")

}
\seealso{
\code{\link{ContractType}} and \code{\link{EventSeries}}
}
