% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YieldCurve.R
\name{MarketInterestRate}
\alias{MarketInterestRate}
\title{\code{MarketInterestRate}}
\usage{
MarketInterestRate(rate, refDate, label, ...)
}
\arguments{
\item{rate}{a numeric to set the constant market rate.}

\item{date}{a character indicating the reference date.}

\item{label}{a character to name the \code{\link{YieldCurve}} object.}
}
\value{
an object of type \code{\link{YieldCurve}}.
}
\description{
Constructor for a \code{\link{YieldCurve}} object with constant
rates for all tenors.
}
\examples{
yc_flat <- MarketInterestRate(0.05, "2015-01-01", "MarketRate")

}
