% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelStructure.R
\name{ModelStructure}
\alias{ModelStructure}
\title{Class that contains the whole model of an enterprise or institution.}
\usage{
ModelStructure(name, ...)
}
\description{
The class is implemented as a data.tree structure. 
It contains a hierarchical model structure.
The upper level is predefined with the nodes "Active", "Passive" and
"Operations".
}
