% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimeBuckets.R
\name{names,timeBuckets-method}
\alias{names,timeBuckets-method}
\title{\code{print} method for class \code{timeBuckets}.}
\usage{
\S4method{names}{timeBuckets}(x)
}
\value{
Labels for time buckets as defined in slot \code{bucketLabs}.
}
\description{
\code{print} method for class \code{timeBuckets}.
}
