% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newbiz.R
\name{newbiz}
\alias{newbiz}
\alias{newbiz,Node,timeDate,matrix,list,RiskFactorConnector-method}
\title{Simulate new (future) business on the PortfolioTree}
\usage{
newbiz(object, by, growth, templates, rf)

\S4method{newbiz}{Node,timeDate,matrix,list,RiskFactorConnector}(object, by, growth, templates, rf)
}
\arguments{
\item{by}{A vector of ordered \link{timeDate}s providing the time-axis along 
which the business simulation is to be conducted.}

\item{rf}{The \link{RiskFactorConnector} providing the market environment under 
which the business scenario is to be evaluated.}

\item{ptf}{A \link{Portfolio} of \link{ContractTypes} which gives the existing
business.}

\item{tree}{A \link{PortfolioTree} with fields \code{leafs}, \code{business}, 
\code{targets} specified that drives the new business generating process.}
}
\value{

}
\description{
A going-concern view on an organization (or \link{PortfolioTree}) includes
the simulation of continuation of the organization's business. This method
allows to compute new future business in terms of ACTUS \link{ContractTypes}
according to a certain model contract and target growth on each "leaf" 
(or terminal account) in the \link{PortfolioTree}.
}
\details{
New business is generated along the time axis defined by method parameter 
\code{by}. At each point in time in \code{by}, the target and actual 
quantity in business, as measured in terms of nominal value, are compared and for the positive 
parts of difference \code{target - actual} new business is generated in form 
of new \link{ContractType}. The new business is then added to the \link{Portfolio}
of existing contracts as provided with method argument \code{ptf}.

The new business generating process is defined by the following information 
elements provided with the \link{PortfolioTree} in method argument \code{tree}:
\itemize{
\item{leafs}{The leafs, or terminal accounts, 
provide the structure for the 
simulation of new business. That is, scenarios for new business can be 
defined only for the leafs of a \link{PortfolioTree}.}
\item{business}{Business contains a model 
contract in terms of a list of
named ACTUS Contract Attributes (as would be used in argument \code{what} 
used with the \link{set}-method applied to a \link{ContractType}). This
model contract is used as a template for all new business contracts 
simulated on this particular leaf.}
\item{targets}{A \code{(number of leafs) x (number of leafs)} 
numeric matrix
containing on the diagonal the growth rates (\code{by}-period wise) on the 
leafs themselves and on the outer-diagonals the spill-over rates from one
leaf to another leaf.}
}

For a demo use \code{demo(NewBusinessSimulation)}.
}
\examples{

}
\seealso{
\link{PortfolioTree}
}
