% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Operations.R
\name{OperationalCF}
\alias{OperationalCF}
\title{OperationalCF Contract class definition}
\usage{
OperationalCF(ContractID, pattern, args, ...)
}
\arguments{
\item{pattern}{A function evaluating the pattern of generated cash flows.}

\item{args}{The list of arguments used when evaluating the pattern}
}
\description{
An operational cash flows contract represents any operational activity 
in monetary units within an organization.
}
\examples{
times = timeSequence(from="2014-01-01", by="3 months", length.out=9)
values = cumsum(c(1,rnorm(8,0.02,0.1)))
idx <- Index(label = "PriceIndex", data = values, charvec = times)

revenue <- function(idx, times) { 
  idx$Data[as.character(times),] * 1000
}
revenue(idx=idx, times=times)
OpCFs <- OperationalCF(
  ContractID="Ops001", Currency="CHF",
  pattern = revenue, 
  args = list( # the argument of the function
    idx = idx,  
   times = as.character(times)))

}
