% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Operations.R
\docType{class}
\name{Operations-class}
\alias{Operations-class}
\alias{OperationalCF-class}
\alias{Investments-class}
\alias{Reserves-class}
\title{Operations Contract class definition}
\arguments{
\item{ContractID}{A unique identifier of the Operations contract}

\item{ContractDealDate}{Timestamp as per when the contract was initiated}

\item{Currency}{The currency in which \link{ReservingPattern}, \link{DepreciationPattern},
\link{IncomePattern} are denominated}

\item{Params}{A list containing parameters used in the \link{CashFlowPattern}, 
\link{InvestPattern}, and \link{ReservePattern} functions}

\item{ReservePattern}{A function evaluating the pattern of building reserves. The 
                      function must implemente two arguments:
\itemize{
\item{model}{The \link{RiskFactorConnector} object used when evaluating the pattern}
\item{params}{The \link{Params} object used when evaluating the pattern}
}}

\item{InvestPattern}{A function evaluating the pattern of investment and depreciation. The 
                      function must implemente two arguments:
\itemize{
\item{model}{The \link{RiskFactorConnector} object used when evaluating the pattern}
\item{params}{The \link{Params} object used when evaluating the pattern}
}}

\item{CashFlowPattern}{A function evaluating the pattern of generated cash flows. 
                       The function must implement two arguments:
\itemize{
\item{model}{The \link{RiskFactorConnector} object used when evaluating the pattern}
\item{params}{The \link{Params} object used when evaluating the pattern}
}}

\item{RiskFactorConnector}{(optional) A risk factor environment within which the Operations
contract will be evaluated}
}
\value{

}
\description{
An operations contract represents any operational activity 
in monetary units within an organization. Such an activity
can be an Investment, Reserving, or any (re-) occuring cost
or income position.
}

