% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Operations.R
\name{Reserves}
\alias{Reserves}
\alias{terms,Operations-method}
\title{Operations Contract constructor definition}
\usage{
Reserves(...)

\S4method{terms}{Operations}(object)
}
\description{
Operations Contract constructor definition
}
