% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ContractPlot.R, R/RiskFactorPlots.R
\docType{methods}
\name{plot}
\alias{plot}
\alias{plot,ContractType,character-method}
\alias{plot,YieldCurve,missing-method}
\alias{plot,RiskFactor,ANY-method}
\title{Generate a graphical representation of an 
\code{\link{EventSeries}}}
\usage{
plot(x, y, ...)

\S4method{plot}{ContractType,character}(x, y, ...)

\S4method{plot}{YieldCurve,missing}(x, y, ...)

\S4method{plot}{RiskFactor,ANY}(x, y, ...)
}
\arguments{
\item{x}{The object carrying the contract events to be potted}

\item{y}{(only if \code{x} is an objecto of class 
\code{\link{ContractType}}) A character giving the analysis
as per which to plot contract events}

\item{...}{Additional graphical parameters.}

\item{object}{The \link{RiskFactor} object to plot}
}
\value{





}
\description{
This method generates a graphical representation of the 
contract events generated by an ACTUS CT.

Create a graphical representation of a \link{YieldCurve}-
object.

Create a graphical representation of an index-like 
\link{RiskFactor}-object.
}
\details{
The graphical representation draws the series of contract 
events and indicates generated cash flows by arrows (incoming 
cash flows are represented by arrows pointing to the x-axis,
outgoing cash flows by arrows pointing away from the x-axis).
A color code indicates what type a certain cash flow relates 
to (cf. legend for further explanaitions). 

\code{plot} may be used with an object of one of classes 
\code{\link{ContractType}} \code{\link{EventSeries}}. If used
with an object of class \code{\link{ContractType}}, the analysis
date as per which events should be drawn has to be provided as 
argument \code{y}.
}
\examples{
pam = Pam()
set(pam, what=list(
                   ContractID = "001",
                   Currency = "CHF",
                   ContractRole = "RPA",
                   StatusDate       = "2012-12-31T00",
                   ContractDealDate = "2012-12-31T00",
                   InitialExchangeDate = "2013-01-01T00",
                   MaturityDate = "2013-03-31T00",
                   NotionalPrincipal = 1000, 
                   NominalInterestRate = 0.01,
                   DayCountConvention = "30E/360"))
plot(pam,"2012-12-31T00") 

yc <- YieldCurve()
tenors <- c("1W", "1M", "6M", "1Y", "2Y", "5Y")
rates <- c(0.001, 0.0015, 0.002, 0.01, 0.02, 0.03)
set(yc, what = list(MarketObjectCode = "YC_Prim",
  Nodes = list(ReferenceDate = "2015-01-01T00", Tenors = tenors, Rates = rates)))
plot(yc)

idx <- Index()
times <- c("2015-01-01T00", "2016-01-01T00", "2017-01-01T00", "2018-01-01T00",
           "2019-01-01T00")
values <- c(100, 110, 120, 130, 140)
set(idx, what=list(
  MarketObjectCode = "IND_CPI_EA",
  Data=list(Dates=times,Values=values)))
plot(idx)

}
\seealso{
\code{\link{ContractType}} and \code{\link{EventSeries}}

\link{YieldCurve},\link{plot}

\link{ReferenceIndex},\link{plot}
}
