% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/presentValue.R
\name{presentValue}
\alias{presentValue}
\title{\code{presentValue}}
\usage{
presentValue(x, yield, yieldCurve, from, isPercentage)
}
\arguments{
\item{x}{a contract type, for which to calculate the NPV.}

\item{yield}{a numeric, indicating the percentage yield used to discount.}

\item{yieldCurve}{an object of type \code{\link{YieldCurve}} or 
\code{\link{DynamicYieldcurve}} to calculate discount 
factors from.}

\item{from}{a character indicating the date as for which the NPV is calculated.}

\item{isPercentage}{a logical, indicating if the 'yield' is inserted as percentage or not.
(default is TRUE).}
}
\value{
a numeric, representing the Net Present Value (NPV) of the contract.
}
\description{
Function which calculates the Net Present Value (NPV) for passed
contract types.
}
\examples{
b <- bond("2013-12-31", maturity = "5 years", nominal = 50000, 
           coupon = 0.02, couponFreq = "1 years")
npv <- presentValue(b, yield = 2) # result: 0 due to same coupon as yield

}
