% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priceVsYield.R
\name{priceVsYieldPlot}
\alias{priceVsYieldPlot}
\title{\code{priceVsYieldPlot}}
\usage{
priceVsYieldPlot(x, prices, isPercentage, from, drawDuration, drawConvexity)
}
\arguments{
\item{x}{a contract type, for which to calculate the YTMs.}

\item{prices}{a numeric vector, indicating the price used for calculating the YTMs.}

\item{isPercentage}{a logical, indicating if the 'price' is passed as percentage or not.
(default is TRUE).}

\item{from}{a character indicating the date as for which the YTM is calculated.}

\item{drawDuration}{a logical indicating if the duration should also be plotted.}

\item{drawConvexity}{a logical indicating if the convexity should also be plotted..}
}
\description{
Function which calculates the yield to maturities (YTMs) for a passed
contract type and a vector of prices.
}
\examples{
b <- bond("2013-12-31", maturity = "5 years", nominal = 50000, 
           coupon = 0.02, couponFreq = "1 years")
priceVsYieldPlot(b, prices = c(80, 90, 100))
priceVsYieldPlot(b, prices = c(80, 90, 100), drawDuration=TRUE, drawConvexity=TRUE)

}
