% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priceVsYield.R
\name{priceVsYieldTable}
\alias{priceVsYieldTable}
\title{\code{priceVsYieldTable}}
\usage{
priceVsYieldTable(x, prices, isPercentage, from)
}
\arguments{
\item{x}{a contract type, for which to calculate the YTMs. This can also be 
a timeSeries or EventSeries object.}

\item{prices}{a numeric vector, indicating the price used for calculating the YTMs.}

\item{isPercentage}{a logical, indicating if the 'price' is passed as percentage or not.
(default is TRUE).}

\item{from}{a character indicating the date as for which the YTM is calculated.}
}
\value{
a data.frame, representing the prices of the contract and the 
        yield to maturities (YTM).
}
\description{
Function which calculates the yield to maturities (YTMs) for a passed
contract type and a vector of prices.
}
\examples{
b <- bond("2013-12-31", maturity = "5 years", nominal = 50000, 
           coupon = 0.02, couponFreq = "1 years")
ytm_tbl <- priceVsYieldTable(b, prices = c(80, 90, 100))

}
