% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimeBuckets.R
\name{print,cfStream-method}
\alias{print,cfStream-method}
\title{\code{print} method for class \code{cfStream}.}
\usage{
\S4method{print}{cfStream}(x, format = "\%d.\%m.\%y", sep = "-")
}
\description{
\code{print} method for class \code{cfStream}.
}
