% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimeBuckets.R
\name{print,timeBuckets-method}
\alias{print,timeBuckets-method}
\title{\code{print} method for class \code{timeBuckets}.}
\usage{
\S4method{print}{timeBuckets}(x)
}
\description{
\code{print} method for class \code{timeBuckets}.
}
