% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Portfolio.R
\docType{class}
\name{Portfolio-class}
\alias{Portfolio-class}
\title{A Reference Class that represents a portfolio of multiple rActus contracts}
\description{
A Portfolio is a simple structure for the representation of 
a collection of \pkg{rActus} \code{ContractType} objects. 
Portfolio allows to compute cash flows and additonal analytical
results for the whole collection of contracts.
}
\section{Fields}{

\describe{
\item{\code{contracts}}{A list for the rActus contracts the portfolio consists of}
}}


\examples{
data(BondPortfolio)
ptf <- Portfolio()
import(ptf,BondPortfolio, valuationEngines=TRUE)
ptf

}
\seealso{
\code{\link{PortfolioFast}}
}
