% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Portfolio.R
\docType{methods}
\name{Portfolio}
\alias{Portfolio}
\alias{Portfolio,ANY-method}
\alias{CTterms}
\alias{ids}
\title{Portfolio-class constructor}
\usage{
Portfolio(...)

\S4method{Portfolio}{ANY}(...)

CTterms(x, i, vars, ...)

ids(x, ...)
}
\arguments{
\item{...}{}
}
\value{
An object of class \code{Portfolio}
}
\description{
Create an instance of \code{Portfolio} class.
}
\examples{
data(BondPortfolio)
ptf <- Portfolio()
import(ptf,BondPortfolio, valuationEngines=TRUE)
ptf

}
\seealso{
\code{\link{PortfolioFast}}
}
