% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RiskFactor.R
\docType{class}
\name{RiskFactor-class}
\alias{RiskFactor-class}
\title{A Reference Class that represents all Risk Factor objects}
\description{
This class is only used in a virtual sense in that it is
not intended to create instances of it. It's use is to 
serve as parent to various implementations of Risk Factor 
types as a means for designing method-inheritage.
}
\section{Fields}{

\describe{
\item{\code{MarketObjectCode}}{character name of the risk factor}

\item{\code{Data}}{data.frame representating time series data}
}}


\seealso{
\code{\link{YieldCurve, ReferenceIndex, ForeignExchangeRate}}
}
