% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RiskFactorConnector.R
\docType{class}
\name{RiskFactorConnector-class}
\alias{RiskFactorConnector-class}
\title{A Reference Class representing the ACTUS concept of a
RiskFactorConnector}
\description{
The cash flows of ACTUS CTs may be 'linked' to (stochastic)
external risk factors such as interest rates or reference
indices. The \code{RiskFactorConnector} serves as the 
interface. It is in fact a collection of risk factors each
with the corresponding 'MarketObjectCode' as it's unique
identifier. 
Instances of R/Java risk factors may be added to the
\code{RiskFactorConnector} which then may be 'linked' to
a given ACTUS CT.
}
\section{Fields}{

\describe{
\item{\code{jref}}{A rJava java object reference}
}}


\seealso{
\code{\link{RiskFactor, YieldCurve, ReferenceIndex}}
}
