% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DynamicYieldCurve.R
\docType{methods}
\name{rates}
\alias{rates}
\alias{rates,}
\alias{DynamicYieldCurve,}
\alias{charachter,}
\alias{missing-method}
\alias{character,}
\alias{character-method}
\alias{rates,YieldCurve,character,missing-method}
\alias{YieldCurve,}
\alias{rates,YieldCurve,character,character-method}
\alias{rates,DynamicYieldCurve,character,missing-method}
\alias{rates,DynamicYieldCurve,character,character-method}
\title{Generic method to retrieve the rate(s) for (a) specific
tenor(s) from a \code{\link{DynamicYieldCurve}} object}
\usage{
rates(object, termEnd, termStart, ...)

\S4method{rates}{YieldCurve,character,missing}(object, termEnd, termStart, isDateEnd = FALSE, refdate = NULL, ...)

\S4method{rates}{YieldCurve,character,character}(object, termEnd, termStart, isDateEnd = FALSE, refdate = NULL, ...)

\S4method{rates}{DynamicYieldCurve,character,missing}(object, termEnd, termStart, isDateEnd = FALSE, refdate = NULL, ...)

\S4method{rates}{DynamicYieldCurve,character,character}(object, termEnd, termStart, isDateEnd = FALSE, refdate = NULL, ...)
}
\arguments{
\item{object}{An object of class \code{DynamicYieldCurve} for 
which to return the yield for a given tenor}

\item{termEnd}{The tenor for which to return its yield. 
Can be a single value or a vector of tenors.}

\item{termStart}{(optional) For the forward rate at t0 
between times t1 and t2, termEnd refers to t2 and
termStart to t1. Is a single value. If combined 
with a vector for termEnd, then termStart remains
the same for all t2 defined in termEnd.}

\item{...}{Additional parameters:
\itemize{
 \item{"isDateEnd"}{logical indicating whether termEnd is 
       of date (TRUE) or term (FALSE) format. Date format is
       'YYYY-MM-DDTXX' with 'XX'=00 for beginning of day, or
       24 for end of day, and term format is 'IP' with 'I' 
       an integer and 'P' indicating the period (D=days, 
       W=weeks, M=months, Q=quarters, H=halfyears, Y=years).
       Default is isDateEnd=FALSE.}
}}
}
\value{
numeric The yield for the defined tenor
}
\description{
A yield curve is a time-structure of yields, i.e. for
different future points in time (tenors) a yield is 
extracted from observed instrument prices. The 
\code{\link{DynamicYieldCurve}} object contains these tenors with
associated yields and allows to retrieve yields for any
tenor by inter-/extrapolation.
}
\examples{
yc <- DynamicYieldCurve()
tenors <- c("1W", "1M", "6M", "1Y", "2Y", "5Y")
rates <- c(0.001, 0.0015, 0.002, 0.01, 0.02, 0.03)
set(yc, what = list(MarketObjectCode = "YC_Prim",
  Nodes = list(ReferenceDate = "2015-01-01T00", Tenors = tenors, Rates = rates)))

rates(yc, "1Y")  # 1-year spot rate
rates(yc, "2016-01-01T00", isDateEnd=TRUE) # again, 1-year spot rate
rates(yc, "1Y", "2015-07-01T00")  # 1-year forward rate at 2015-07-01T00

}
\seealso{
\code{\link{discountFactors}}
}
