% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ContractModel.R, R/ContractType.R,
%   R/YieldCurve.R, R/Portfolio.R, R/EventSeries.R, R/YieldCurve2.R,
%   R/ForeignExchangeRate.R, R/MultiCurrencyDiscountingEngine.R,
%   R/PortfolioTree.R, R/ReferenceIndex.R
\docType{methods}
\name{set}
\alias{set}
\alias{set,ContractType,list-method}
\alias{set,YieldCurve,list-method}
\alias{set,ReferenceIndex,list-method}
\alias{set,ForeignExchangeRate,list-method}
\alias{set,ValuationEngineModel,list-method}
\alias{set,jobjRef,list-method}
\alias{set,ContractType,ValuationEngine-method}
\alias{set,ContractType,RiskFactorConnector-method}
\alias{set,ContractType,ContractModel-method}
\alias{set,Portfolio,list-method}
\alias{set,Portfolio,RiskFactorConnector-method}
\alias{set,EventSeries,list-method}
\alias{set,YieldCurve2,list-method}
\alias{set,MultiCurrencyDiscountingModel,list-method}
\alias{set,PortfolioTree,list-method}
\title{Generic method to set the value of term(s) or parameter(s) 
of an ACTUS object}
\usage{
set(object, what, ...)

\S4method{set}{ContractType,ValuationEngine}(object, what)

\S4method{set}{ContractType,list}(object, what)

\S4method{set}{ContractType,RiskFactorConnector}(object, what)

\S4method{set}{YieldCurve,list}(object, what, ...)

\S4method{set}{Portfolio,list}(object, what)

\S4method{set}{Portfolio,RiskFactorConnector}(object, what, ...)

\S4method{set}{EventSeries,list}(object, what)

\S4method{set}{YieldCurve2,list}(object, what, ...)

\S4method{set}{ForeignExchangeRate,list}(object, what, ...)

\S4method{set}{MultiCurrencyDiscountingModel,list}(object, what, ...)

\S4method{set}{PortfolioTree,list}(object, what)

\S4method{set}{ReferenceIndex,list}(object, what, ...)
}
\arguments{
\item{object}{A \code{PortfolioTree}-object}

\item{what}{A list with names and values of fields to change}

\item{...}{}
}
\value{





}
\description{
Implemented e.g. for \code{\link{ContractModel}} where 
\code{set} populates the value for specified contract 
term(s) or for \code{\link{RiskFactor}} implementations, 
\code{\link{ValuationModel}}, and whatever object of this 
package that is carrying parameters.
Note that argument \code{what} has to refer to the exact
names of the terms or parameters to populate. The values
need to be of the respective data type, i.e. numeric,
character, logical, etc. depending on the term to be set.

Allows to change certain elements of a portfolio. In particular, 
this is a convenience function to change the 
Reference Class's field values. Further, the method allows to
assign a \code{RiskFactorConnector} to the portfolio, i.e.
all contracts in the portfolio.
See also the respective documentation in \pkg{rActus}.

A convenience-wrapper to change values of the fields of a
\code{PortfolioTree}.
}
\examples{
pam = Pam()
terms(pam) # get all term names
set(pam, what=list(NotionalPrincipal=1000,
                   DayCountConvention="A/AISDA",
                   InitialExchangeDate="2015-01-01T00"))
get(pam, what=c("NotionalPrincipal", "DayCountConvention", 
                   "InitialExchangeDate"))

# define analysis data
ad <- "2015-01-02T00"

# construct portfolio
data(BondPortfolio)
ptf <- Portfolio()
import(ptf,BondPortfolio, valuationEngines=TRUE)

# construct market model
yc <- YieldCurve()
tenors <- c("1W", "1M", "6M", "1Y", "2Y", "5Y")
rates <- c(0.001, 0.0015, 0.002, 0.01, 0.02, 0.03)
set(yc, what = list(
MarketObjectCode = "YC_EA_AAA",
Nodes = list(ReferenceDate = ad, 
Tenors = tenors, Rates = rates)))

cpi <- Index()
times <- c("2015-01-01T00", "2016-01-01T00", "2017-01-01T00", "2018-01-01T00",
"2019-01-01T00")
values <- c(100, 110, 120, 130, 140)
set(cpi, what=list(
MarketObjectCode = "IND_CPI_EA",
Data=list(Dates=times,Values=values)))

rf <- RFConn()
add(rf, list(yc, cpi))

# assign market model to portfolio
set(ptf, rf, valuationEngines=TRUE)

tree=Tree()
set(tree, what=list(leafs=list(L1=c(1001,1002),L2=c(1003,1004))))

}
\seealso{
\code{\link{terms,get}}

\code{\link{get}}

\link{get}
}
