% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimeBuckets.R
\name{show,cfStream-method}
\alias{show,cfStream-method}
\title{\code{show} method for class \code{cfStream}.}
\usage{
\S4method{show}{cfStream}(object)
}
\description{
This method is defined by means of the print method because otherwise 
the slots "partition" and "currency" are displayed in an ugly way 
through the default method of "show" for "data.frame".
}
