% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EventSeries.R
\docType{methods}
\name{show,EventSeries-method}
\alias{show,EventSeries-method}
\title{\code{EventSeries}-show method}
\usage{
\S4method{show}{EventSeries}(object)
}
\arguments{
\item{object}{An object of class \code{EventSeries}}
}
\value{
show returns an invisible NULL.
}
\description{
displays an \code{EventSeries} in terminal
}
\details{
This method displays an object of class \code{EventSeries} in an formatted way.

Each event's EventDate, EventType, EventValue, and state
variables from the java-object to R (through rJava) and
structures the data in an R-data.frame object.
}
\examples{
pam = Pam()
set(pam, what=list(
                   ContractID = "001",
                   Currency = "CHF",
                   ContractRole = "RPA",
                   StatusDate       = "2012-12-31T00",
                   ContractDealDate = "2012-12-31T00",
                   InitialExchangeDate = "2013-01-01T00",
                   MaturityDate = "2013-03-31T00",
                   NotionalPrincipal = 1000,
                   NominalInterestRate = 0.01,
                   DayCountConvention = "30E/360"))
ad = AD0("2012-12-31T24")
events(pam, ad)


}
\seealso{
\code{\link{ContractType, events}}
}
