% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimeBuckets.R
\name{show,timeBuckets-method}
\alias{show,timeBuckets-method}
\title{\code{show} method for class \code{timeBuckets}.}
\usage{
\S4method{show}{timeBuckets}(object)
}
\description{
\code{show} method for class \code{timeBuckets}.
}
