% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulationManager.R
\name{simulate,SimulationManager,timeDate,timeDate,character-method}
\alias{simulate,SimulationManager,timeDate,timeDate,character-method}
\title{Carries out the simulation.
For every Object, in creates an EventSeries that is the basis for all further
analytics.}
\usage{
\S4method{simulate}{SimulationManager,timeDate,timeDate,character}(object, start, end, by, ...)
}
\description{
Carries out the simulation.
For every Object, in creates an EventSeries that is the basis for all further
analytics.
}
