% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ContractType.R, R/EventSeries.R
\docType{methods}
\name{[,ContractType,character,missing-method}
\alias{[,ContractType,character,missing-method}
\alias{[,ContractType,numeric,missing-method}
\alias{[,ContractType,logical,missing-method}
\alias{[<-,ContractType,character,missing,ANY-method}
\alias{[<-,ContractType,numeric,missing,ANY-method}
\alias{[,EventSeries,character,missing-method}
\alias{[,EventSeries,numeric,missing-method}
\alias{[,EventSeries,numeric,numeric-method}
\alias{[,EventSeries,numeric,logical-method}
\alias{[,EventSeries,numeric,character-method}
\title{\code{ContractType}-subscript methods}
\usage{
\S4method{[}{ContractType,character,missing}(x, i)

\S4method{[}{ContractType,numeric,missing}(x, i)

\S4method{[}{ContractType,logical,missing}(x, i)

\S4method{[}{ContractType,character,missing,ANY}(x, i) <- value

\S4method{[}{ContractType,numeric,missing,ANY}(x, i) <- value

\S4method{[}{EventSeries,character,missing}(x, i)

\S4method{[}{EventSeries,numeric,missing}(x, i)

\S4method{[}{EventSeries,numeric,numeric}(x, i, j)

\S4method{[}{EventSeries,numeric,logical}(x, i, j)

\S4method{[}{EventSeries,numeric,character}(x, i, j)
}
\description{
accesses elements of an \code{ContractType}

accesses elements of an \code{EventSeries}
}
\details{
This method accesses elements of an object of class \code{ContractType}

This method prints an object of class \code{EventSeries} in an formatted way.
Formatting options are provided

each event's EventDate, EventType, EventValue, and state
variables from the java-object to R (through rJava) and
structures the data in an R-data.frame object.
}
