% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimeBuckets.R
\docType{class}
\name{timeBuckets-class}
\alias{timeBuckets-class}
\title{An S4 class that defines a time partition.}
\description{
The partition is defined by a vector of class \code{timeDate} defining as 
sequence \eqn{t = t_1,\cdots,t_{n+1}}. 
The time intervals \eqn{\Delta_i} are left-open and right-closed, 
\eqn{\Delta_i = (t_1,t_{i+1}]}, \eqn{i = 1,\cdots,n}
}
\section{Slots}{

\describe{
\item{\code{timeDate}}{Defines the sequence \eqn{\{t_i\}}, \eqn{i = 1,\cdots,n}}

\item{\code{bucketLabs}}{\code{character} vector containing labels for the time 
intervals.}

\item{\code{breakLabs}}{\code{character} vector containing labels for \eqn{\{t_i\}}.}
}}

