% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimeBuckets.R
\docType{methods}
\name{timeBuckets}
\alias{timeBuckets}
\alias{length,timeBuckets-method}
\alias{cfStream-class}
\alias{cfStream}
\alias{periodNames}
\alias{aDates2tDates}
\title{Constructor for class \code{timeBuckets}.}
\usage{
timeBuckets(by, bucketLabs = NULL, breakLabs = NULL, ...)

\S4method{length}{timeBuckets}(x)

cfStream(buckets, currency = "None", data)

periodNames(by, format = "\%d.\%m.\%y", sep = "-")

aDates2tDates(char)
}
\arguments{
\item{by}{\code{character} vector of dates.}

\item{bucketLabs}{Chararacter vector containing the labels of the time 
buckets.}

\item{breakLabs}{Chracter vector containing the labels of the break points.}

\item{x}{Object of class \code{timeBuckets}.}

\item{buckets}{\code{timeBuckets} object containing the time intervals.}

\item{currency}{\code{character} with currency code.}

\item{data}{Cash flow data. Can be \code{numeric}, \code{integer}, or 
\code{data.frame}.}

\item{format}{\code{character} containing the formatting information for 
the "from" and "to" part of the output.}

\item{sep}{Separator of "from" and "to" part of output.}

\item{char}{Character string in ACTUS time format.}
}
\value{
Object of class \code{timeBuckets}.

\code{integer} containing the number of time intervals in \code{x}.

\code{cfStream} object.

\code{character} vector with labels.

\code{timeDate} object.
}
\description{
S4 class that contains cash flows occuring durring the time intervals defined
by an \code{timeBuckets} object.

The labels have the form "from%sep%to".
}
\details{
This class extend \code{data.frame}.
}
\section{Slots}{

\describe{
\item{\code{partition}}{\code{timeBuckets} object.}

\item{\code{currency}}{\code{character} with currency code.}
}}

