% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PortfolioTree.R
\docType{class}
\name{PortfolioTree-class}
\alias{PortfolioTree-class}
\title{A hierarchical tree structure for the organization of a portfolio}
\description{
This Reference Class represents a hierarchical tree structure
that allows organizing a portfolio of contracts into sub-portfolios
as is the case e.g. for an organization's balance sheet.
}
\details{
A tree can then be passed to methods \link{value}, \link{liquidity}, \link{income} 
by specifying the optional argument \code{tree=}. If a tree is specified, these
methods will then compute and return the respective quantities on the various
aggregate levels in the tree's hierarchy rather than the granular contract level.
}
\section{Fields}{

\describe{
\item{\code{branches}}{a list of intermediate sub-portfolios each having a name and
a vector indicating the names of its sub-portfolios.}

\item{\code{leafs}}{a list of terminal sub-portfolios to which the contracts 
of the \code{Portfolio} are assigned. Terminal sub-portfolio here means that
a \code{leaf} can itself not have sub-portfolios but only contracts assigned
to it. Hence, a leaf has a name and a vector indicating the \code{ContractID}s
of its contracts.}
}}


\examples{
tree=Tree(list(
               branches=list(B0=c("B1","B2"),
                             B1="L1",
                             B2=c("L2","L3")),
               leafs=list(L1=c("106","107"),
                          L2=c("108","109"),
                          L3=c("110","111"))
                ))
tree

}
\seealso{
\link{Portfolio}, \link{ContractType}
}
