% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PortfolioTree.R
\name{Tree}
\alias{Tree}
\alias{Tree,ANY-method}
\title{Create a new \code{PortfolioTree}}
\usage{
Tree(...)

\S4method{Tree}{ANY}(...)
}
\arguments{
\item{...}{(optional) Comma-separated elements (or a list of elements) specifying field values to be attached to the new object}
}
\value{
A new \code{PortfolioTree}-object
}
\description{
This method creates a new \code{PortfolioTree}-object.
}
\examples{
tree=Tree(list(
               branches=list(B0=c("B1","B2"),
                             B1="L1",
                             B2=c("L2","L3")),
               leafs=list(L1=c("106","107"),
                          L2=c("108","109"),
                          L3=c("110","111"))
                ))
tree

}
\seealso{

}
