% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ContractModel.R, R/ValuationEngine.R,
%   R/ContractType.R
\docType{methods}
\name{terms}
\alias{terms}
\alias{terms,ContractType-method}
\alias{terms,ValuationEngine-method}
\alias{terms,ContractModel-method}
\alias{terms,jobjRef-method}
\alias{terms,ValuationEngineModel-method}
\title{Generic method to return existing terms or parameters of an
ACTUS object}
\usage{
terms(object)

\S4method{terms}{ContractModel}(object)

\S4method{terms}{ValuationEngine}(object)

\S4method{terms}{ContractType}(object)
}
\arguments{
\item{object}{The object for which to return terms}
}
\value{
character A character vector of the names of 
       available terms
}
\description{
Implemented e.g. for ACTUS contract models where \code{terms}
returns all available contract terms or for valuation 
engine models where the method returns the valuation model
parameters.
}
\examples{
pam = Pam()
terms(pam) # returns all Principal At Maturity terms

dc = DcModel()
terms(dc) # returns all parameters of the discounting model

}
\seealso{
\code{\link{get,set}}
}
