% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Value.R, R/CurrentAccount.R,
%   R/ModelStructure.R, R/Operations.R
\docType{methods}
\name{value}
\alias{value}
\alias{value,ContractType,character,character,missing-method}
\alias{value,ContractType,AD0,character,missing-method}
\alias{value,ContractType,timeDate,character,missing-method}
\alias{value,ContractType,timeBuckets,character,missing-method}
\alias{value,ContractType,timeBuckets,character,YieldCurve-method}
\alias{value,ContractType,character,character,ValuationEngine-method}
\alias{value,ContractType,timeDate,character,ValuationEngine-method}
\alias{value,ContractType,timeBuckets,character,ValuationEngine-method}
\alias{value,ContractType,AD0,character,ValuationEngine-method}
\alias{value,EventSeries,AD0,character,missing-method}
\alias{value,EventSeries,timeDate,character,missing-method}
\alias{value,EventSeries,timeBuckets,character,missing-method}
\alias{value,EventSeries,timeDate,character,DiscountingEngine-method}
\alias{value,EventSeries,timeBuckets,character,DiscountingEngine-method}
\alias{value,eventList,timeDate,character,missing-method}
\alias{value,eventList,timeBuckets,character,missing-method}
\alias{value,eventList,timeDate,character,DiscountingEngine-method}
\alias{value,eventList,timeBuckets,character,DiscountingEngine-method}
\alias{value,EventSeries,character,character,DiscountingEngine-method}
\alias{value,EventSeries,character,character,missing-method}
\alias{value,eventList,character,character,missing-method}
\alias{value,eventList,character,character,DiscountingEngine-method}
\alias{value,eventList,character,character,list-method}
\alias{value,CurrentAccount,character,character,missing-method}
\alias{value,CurrentAccount,character,character,ValuationEngine-method}
\alias{value,Node,timeBuckets,character,ANY-method}
\alias{value,Operations,AD0,character,missing-method}
\alias{value,Operations,character,character,missing-method}
\alias{value,Operations,timeDate,character,missing-method}
\alias{value,Operations,timeBuckets,character,missing-method}
\alias{value,Operations,character,character,DiscountingEngine-method}
\alias{value,Operations,timeDate,character,DiscountingEngine-method}
\alias{value,Operations,timeBuckets,character,DiscountingEngine-method}
\title{Derive the value at AD0 of a contract}
\usage{
value(object, by, type, method, ...)

\S4method{value}{ContractType,character,character,missing}(object, by, type, method, ...)

\S4method{value}{ContractType,AD0,character,missing}(object, by, type, method, ...)

\S4method{value}{ContractType,timeDate,character,missing}(object, by, type, method, ...)

\S4method{value}{ContractType,timeBuckets,character,missing}(object, by, type, method, ...)

\S4method{value}{ContractType,timeBuckets,character,YieldCurve}(object, by, type, method, spread = 0, ...)

\S4method{value}{ContractType,character,character,ValuationEngine}(object, by, type, method, ...)

\S4method{value}{ContractType,timeDate,character,ValuationEngine}(object, by, type, method, ...)

\S4method{value}{ContractType,timeBuckets,character,ValuationEngine}(object, by, type, method, ...)

\S4method{value}{ContractType,AD0,character,ValuationEngine}(object, by, type, method, ...)

\S4method{value}{EventSeries,AD0,character,missing}(object, by, type, method, ...)

\S4method{value}{EventSeries,timeDate,character,missing}(object, by, type, method, ...)

\S4method{value}{EventSeries,timeBuckets,character,missing}(object, by, type, method, ...)

\S4method{value}{EventSeries,timeDate,character,DiscountingEngine}(object, by, type, method, ...)

\S4method{value}{EventSeries,timeBuckets,character,DiscountingEngine}(object, by, type, method, ...)

\S4method{value}{eventList,timeDate,character,missing}(object, by, type, method, ...)

\S4method{value}{eventList,timeBuckets,character,missing}(object, by, type, method, ...)

\S4method{value}{eventList,timeDate,character,DiscountingEngine}(object, by, type, method, ...)

\S4method{value}{eventList,timeBuckets,character,DiscountingEngine}(object, by, type, method, ...)

\S4method{value}{EventSeries,character,character,DiscountingEngine}(object, by, type, method, digits = 2, ...)

\S4method{value}{EventSeries,character,character,missing}(object, by, type, digits = 2, ...)

\S4method{value}{eventList,character,character,missing}(object, by, type, method, ...)

\S4method{value}{eventList,character,character,DiscountingEngine}(object, by, type, method, ...)

\S4method{value}{eventList,character,character,list}(object, by, type, method, ...)

\S4method{value}{CurrentAccount,character,character,missing}(object, by, type, method, end_date, ...)

\S4method{value}{CurrentAccount,character,character,ValuationEngine}(object, by, type, method, end_date, ...)

\S4method{value}{Node,timeBuckets,character,ANY}(object, by, type)

\S4method{value}{Operations,AD0,character,missing}(object, by, type, method, ...)

\S4method{value}{Operations,character,character,missing}(object, by, type, method, ...)

\S4method{value}{Operations,timeDate,character,missing}(object, by, type, method, ...)

\S4method{value}{Operations,timeBuckets,character,missing}(object, by, type, method, ...)

\S4method{value}{Operations,character,character,DiscountingEngine}(object, by, type, method, ...)

\S4method{value}{Operations,timeDate,character,DiscountingEngine}(object, by, type, method, ...)

\S4method{value}{Operations,timeBuckets,character,DiscountingEngine}(object, by, type, method, ...)
}
\arguments{
\item{object}{object The \code{ContractType} or \code{EventSeries}-object for which to derive the value}

\item{by}{time as per which to compute the value}

\item{type}{A character representing the type of value (either 'nominal' or 'markToModel')}

\item{method}{(optional) for type='markToModel' a 'ValuationEngine' can be specified according to which the value is computed}

\item{...}{Currently unused}
}
\value{
A \code{numeric} object representing the contract's value
}
\description{
Different value-concepts can be derived from a financial instrument
or the resulting EventSeries, respectively. Currentently, these are
Nominal value and Mark-to-Model value.

#' @export
#' @include TimeBuckets.R
#' @rdname inc-methods
setMethod(f = "income", signature = c("Operations", "timeBuckets", "missing"),
          definition = function(object, by, type, ...){
            inc <- income(object, as.timeDate(by), type="marginal", ...)
            names(inc) <- by@bucketLabs
            return(inc)
          })
}
\details{
Nominal value is the value of a Maturity instrument that is currently 
outstanding as a dept from the counterparty on the liability side to
the party on the asset side. The payment pattern of the nominal value
over time is what distincts different Maturity instrument types from
each other.
Actus CT algorithms generate the nominal value over time together with
cash flows. Hence, the nominal value for a certain analysis date may 
only be derived from the first-level results of a Maturity contract.

Mark-to-model value is the value of a financial instrument according
to a certain valuation model. E.g. for Maturity instruments, this is usually
its net-present-value. Mostly, this value is used if no value can be observed 
for this instrument on the market. This can be the case if the market for this
instrument is illiquid. 
In order for a mark-to-model value to be derived, the CT algorithm must have
been processed and an appropriate \code{ValuationEngine} either assigned in 
advance or past as method argument 'method'.


}
\examples{
pam <- Pam()
set(pam, what=list(
                 ContractID = "001",
                 Currency = "CHF",
                 Calendar = "Weekday",
                 ContractRole = "RPA",               
                 StatusDate       = "2016-05-30T00",
                 ContractDealDate = "2016-05-30T00",
                 InitialExchangeDate = "2016-05-30T00",
                 MaturityDate = "2016-09-01T00",
                 NotionalPrincipal = 1000,
                 NominalInterestRate = 0.00, 
                 PremiumDiscountAtIED = -100,
                 DayCountConvention = "30E/360",
                 BusinessDayConvention = "SCF"))
ad <- "2016-06-01T00"
yc <- YieldCurve()
set(yc, what=list(Nodes=list(ReferenceDate=ad,
                             Tenors=c("1M","10Y"),
                             Rates=c(0.005,0.02)),
                  MarketObjectCode = "RiskFreeCurve"))
rf <- RFConn()
add(rf,yc)
set(pam,rf)

# derive nominal value
value(pam,by=ad,type="nominal")

# define valuation engine for mark-to-model value
dcEngine <- DcEngine()
set(dcEngine,list(RiskFactorObjectLink="RiskFreeCurve",
                  DiscountingSpread=0.0))
set(dcEngine,rf)
set(pam,dcEngine)
value(pam,by=ad,type="markToModel")

# again with a different engine
set(dcEngine,list(RiskFactorObjectLink="RiskFreeCurve",
                  DiscountingSpread=0.1))
value(pam,by=ad,type="markToModel",method=dcEngine)

}
\seealso{
\code{\link{ContractType}} and \code{\link{ValuationEngine}}
}
