% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ValuationEngine.R
\docType{class}
\name{ValuationEngine-class}
\alias{ValuationEngine-class}
\title{A Reference Class that represents all Valuation Engine objects}
\description{
This class is only used in a virtual sense in that it is
not intended to create instances of it. It's use is to 
serve as parent to various implementations of Valuation
Engine types as a means for designing method-inheritage.
}
\section{Fields}{

\describe{
\item{\code{jref}}{A rJava java object reference}
}}


\seealso{
\code{\link{ValuationEngineModel, DiscountingEngine, 
               CAPMEngine, MultiCurrencyDiscountingEngine}}
}
