% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ValuationEngine.R, R/DiscountingEngine.R,
%   R/CapmEngine.R, R/MultiCurrencyDiscountingEngine.R
\docType{methods}
\name{VE}
\alias{VE}
\alias{VE,jobjRef-method}
\alias{VE,character-method}
\alias{DcEngine}
\alias{DcEngine,missing-method}
\alias{DcEngine,RiskFactorConnector-method}
\alias{CapmEngine}
\alias{CapmEngine,ANY-method}
\alias{XfxDcEngine}
\alias{XfxDcEngine,ANY-method}
\title{\code{ValuationEngine}-class constructor}
\usage{
VE(object)

\S4method{VE}{character}(object)

DcEngine(rf, ...)

\S4method{DcEngine}{missing}(rf, ...)

\S4method{DcEngine}{RiskFactorConnector}(rf, ...)

CapmEngine(...)

\S4method{CapmEngine}{ANY}(...)

XfxDcEngine(...)

\S4method{XfxDcEngine}{ANY}(...)
}
\arguments{
\item{object}{If an object of class \code{jobjRef}, then 
this parameter is expected to be a reference to a Java
valuation engine and a new R-object of the same class is 
created where the Java reference is attached as the
classes' \code{jref}-field. If a character, then
\code{object} is expected to be the R-class name of
the valuation engine to be instantiated.}
}
\value{
An object of a class extending \code{ValuationEngine}
}
\description{
Create an instance of an implementation of class 
\code{ValuationEngine} (e.g. \code{\link{DiscountingEngine}},
etc). 
This constructor is in fact a short cut to the constructors
of the implemented classes such as \code{\link{DcEngine}} for 
\code{\link{DiscountingEngine}}. Note that it is not possible 
to instanciate class \code{ValuationEngine} itself but only 
the implementing classes extending \code{ValuationEngine}.
}
\examples{
# example 1: create a new 'DiscountingEngine' object
dc = VE("DiscountingEngine")

# example 2: attach the reference to a Java 'DiscountingEngine'
#            object to a new R-'DiscountingEngine' object. Note,
#            the new object will refer to the same Java engine.
dc = VE("DiscountingEngine")
set(dc, what=list(
                   RiskFactorObjectLink="YC_CHF",
                   dc.spread=0.05))
# same.dc = VE(dc$jref)   # This command doesn't work

}
\seealso{
\code{\link{VEM, DcEng}}
}
