% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ValuationEngine.R
\docType{class}
\name{ValuationEngineModel-class}
\alias{ValuationEngineModel-class}
\title{A Reference Class that represents all ValuationEngine
Models objects}
\description{
This class is only used in a virtual sense in that it is
not intended to create instances of it. It's use is to 
provide the class structure, i.e. field definitions, to 
various implementations of Contract Model types and as a 
means for designing method-inheritage.
}
\details{
Generally, a Valuation Engine Model has one field that 
contains a reference to a java object of the same class in 
the java ACTUS library (e.g. for 
\code{\link{DiscountingModel}} the referenced 
object will be of class 
org.actus.misc.valuation.DiscountingModel). Note that
the object "carrying" the attribute values, i.e. the
ValuationEngineModel object, lives in Java, or the JVM 
respectively, and not in R. Parameter values may be 
accessed/changed using methods 
\code{\link{get}}/\code{\link{set}} which uses the JNI
interface through R-package rJava.

For a list of available engine attributes use 
method \code{\link{terms}}.
}
\section{Fields}{

\describe{
\item{\code{jref}}{A rJava java object reference}
}}


\seealso{
\code{\link{ValuationEngine, 
DiscountingEngineModel}}
}
