% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yearFraction.R
\name{yearFraction}
\alias{yearFraction}
\alias{yearFraction,}
\alias{charachter,}
\alias{charachter-method}
\title{Generic function to calculate year fractions}
\usage{
yearFraction(start_dates, end_dates, ...)
}
\arguments{
\item{start_dates}{Array of dates from which to 
calculate year fractions in format "yyyy-mm-dd".}

\item{end_dates}{Array of dates until which to 
calculate year fractions in format "yyyy-mm-dd".}

\item{...}{Currently unused}
}
\value{
numeric Array of year fraction calculated from 
        date vectors
}
\description{
This function takes two date arrays and an optional input
for the daycount convention and calculates year fractions
accordingly.
}
\examples{
yearFraction("2020-01-01","2020-01-31","30E/360")
yearFraction(c("2020-01-01","2020-01-31"),"2020-01-31")
yearFraction("2020-01-01",c("2020-01-01","2020-01-31"))
yearFraction(c("2020-01-01","2020-01-31"),c("2020-01-31","2020-02-28"))

}
