% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yieldToMaturity.R
\name{yield}
\alias{yield}
\title{\code{yield}}
\usage{
yield(x, price, isPercentage, from)
}
\arguments{
\item{x}{a contract type, for which to calculate the YTM.}

\item{price}{a numeric, indicating the price used for calculating the YTM.}

\item{isPercentage}{a logical, indicating if the 'price' is inserted as percentage or not.
(default is TRUE).}

\item{from}{a character indicating the date as for which the YTM is calculated.}
}
\value{
a numeric, representing the yield to maturity (YTM) of the contract.
}
\description{
Function which calculates the yield to maturity (YTM) for passed
contract types.
}
\examples{
b <- bond("2013-12-31", maturity = "5 years", nominal = 50000, 
           coupon = 0.02, couponFreq = "1 years")
ytm <- yield(b, price = 100) 

}
