% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DynamicYieldCurve.R
\docType{methods}
\name{discountFactors}
\alias{discountFactors}
\alias{discountFactors,YieldCurve,character,missing-method}
\alias{discountFactors,}
\alias{YieldCurve,}
\alias{character,}
\alias{character-method}
\alias{discountFactors,YieldCurve,character,character-method}
\alias{discountFactors,DynamicYieldCurve,character,missing-method}
\alias{DynamicYieldCurve,}
\alias{discountFactors,numeric,character,character-method}
\alias{discountFactors,DynamicYieldCurve,character,character-method}
\alias{missing-method}
\title{Generic method to retrieve discount factors for specific
tenor(s) from a \code{\link{YieldCurve}} or \code{\link{DynamicYieldCurve}} 
object.}
\usage{
discountFactors(object, end, start, ...)

\S4method{discountFactors}{YieldCurve,character,missing}(
  object,
  end,
  start,
  method = "continuous",
  period = "Y",
  refdate = NULL,
  ...
)

\S4method{discountFactors}{YieldCurve,character,character}(
  object,
  end,
  start,
  method = "continuous",
  period = "Y",
  refdate = NULL,
  ...
)

\S4method{discountFactors}{DynamicYieldCurve,character,missing}(
  object,
  end,
  start,
  method = "continuous",
  period = "Y",
  refdate = NULL,
  ...
)

\S4method{discountFactors}{numeric,character,character}(
  object,
  end,
  start,
  method = "continuous",
  period = "Y",
  refdate = NULL,
  ...
)

\S4method{discountFactors}{DynamicYieldCurve,character,character}(
  object,
  end,
  start,
  method = "continuous",
  period = "Y",
  refdate = NULL,
  ...
)
}
\arguments{
\item{object}{An object of class \code{\link{YieldCurve}} or 
\code{DynamicYieldCurve} from which to extract the 
discount factors.}

\item{end}{character reflecting the discounting period end date.
Can be a single value or a vector of tenors.}

\item{start}{(optional) a character defining the discounting 
period start date. If it is a single value combined 
with a vector for 'end', then 'start' remains
the same for all dates defined in 'end'.}

\item{...}{Additional parameters to be passed.
\itemize{
 \item{"isDateEnd"}{logical indicating whether 'end' is 
       of date (TRUE) or term (FALSE) format. Date format is
       'YYYY-MM-DD'. Default is isDateEnd = FALSE.}
}}
}
\value{
numeric vector of discount factors for the defined periods.
}
\description{
Discount factors for specified tenors are extracted from a 
\code{\link{YieldCurve}} or \code{\link{DynamicYieldCurve}} 
object according to the compounding method specified.
}
\examples{
tenors <- c("1W", "1M", "6M", "1Y", "2Y", "5Y")
rates <- c(0.001, 0.0015, 0.002, 0.01, 0.02, 0.03)
yc <- DynamicYieldCurve(label = "YC_Prim",
                        ReferenceDate = "2015-01-01",
                        Tenors = tenors,
                        Rates = rates)

discountFactors(yc, "1Y")
discountFactors(yc, "2016-01-01", isDateEnd = TRUE)
discountFactors(yc, "1Y", "2015-07-01")

}
\seealso{
\code{\link{rates}}
}
