% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DynamicYieldCurve.R
\docType{methods}
\name{discountFactors}
\alias{discountFactors}
\alias{discountFactors,YieldCurve-method}
\alias{discountFactors,}
\alias{YieldCurve,}
\alias{character,}
\alias{character-method}
\alias{discountFactors,numeric-method}
\alias{DynamicYieldCurve,}
\alias{discountFactors,DynamicYieldCurve-method}
\alias{missing-method}
\title{Generic method to retrieve discount factors for specific
tenor(s) from a \code{\link{YieldCurve}} or \code{\link{DynamicYieldCurve}} 
object.}
\usage{
discountFactors(object, to, from, method = "continuous", 
period = "Y", refdate = NULL, ...)

\S4method{discountFactors}{YieldCurve}(object, ...)

\S4method{discountFactors}{numeric}(object, to, from, ...)

\S4method{discountFactors}{DynamicYieldCurve}(
  object,
  to,
  from,
  by,
  method = "continuous",
  period = "Y",
  ad = NULL,
  refdate,
  ...
)
}
\arguments{
\item{object}{An object of class \code{\link{YieldCurve}} or 
\code{DynamicYieldCurve} from which to extract the 
discount factors.}

\item{...}{Additional parameters to be passed.
\itemize{
 \item{"isDateEnd"} {(deprecated)logical indicating whether 'end' is 
       of date (TRUE) or term (FALSE) format. Date format is
       'YYYY-MM-DD'. Default is isDateEnd = FALSE.}
}}

\item{to}{character reflecting the discounting period end date.
Can be a single value or a vector of tenors.}

\item{from}{(optional) a character defining the discounting 
period start date. If it is a single value combined 
with a vector for 'end', then 'start' remains
the same for all dates defined in 'end'.}

\item{by}{(optional)}

\item{method}{}

\item{period}{}

\item{ad}{}

\item{refdate}{}
}
\value{
numeric vector of discount factors for the defined periods.
}
\description{
Discount factors for specified tenors are extracted from a 
\code{\link{YieldCurve}} or \code{\link{DynamicYieldCurve}} 
object according to the compounding method specified.
}
\examples{
tenors <- c("1W", "1M", "6M", "1Y", "2Y", "5Y")
rates <- c(0.001, 0.0015, 0.002, 0.01, 0.02, 0.03)
yc <- DynamicYieldCurve(label = "YC_Prim",
                        ReferenceDate = "2015-01-01",
                        Tenors = tenors,
                        Rates = rates)

discountFactors(yc, by="1Y")
discountFactors(yc, to="2016-01-01", isDateEnd = TRUE)
discountFactors(yc, by="1Y", from="2015-07-01")

}
\seealso{
\code{\link{rates}}
}
