% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duration.R
\name{duration}
\alias{duration}
\title{\code{duration}}
\usage{
duration(x, type, yield, yieldCurve, price, isPercentage, from)
}
\arguments{
\item{x}{the contract or portfolio of contracts for which to calculate 
the duration.}

\item{type}{a character defining the type of duration; possible types are 
'gen', 'mac' and 'dol'. (default is 'gen').}

\item{yield}{a numeric, indicating the percentage yield used to calculate duration.}

\item{yieldCurve}{Object of type \code{YieldCurve} or 
\code{DynamicYieldCurve} for retrieving discount factors.}

\item{price}{a numeric, indicating the price used for calculating the yield to maturity
of each contract.}

\item{isPercentage}{a logical, indicating if the 'yield' is passed as percentage 
(TRUE) or as fraction (FALSE) (default is TRUE).}

\item{from}{a character indicating the date as for which the net present value
is calculated.}
}
\value{
the duration of the contract or portfolio.
}
\description{
Function which calculates the duration of a portfolio or single contracts.
}
\examples{
bnd1=bond(start="2015-01-01", maturity="30 years", nominal=1000, 
          coupon=0.06, couponFreq="1 year", role="long", variable=FALSE)
duration(bnd1, type="mac", yield=9)

}
