% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Events.R
\docType{methods}
\name{events}
\alias{events}
\alias{events,ContractType,character,missing-method}
\alias{events,ContractType,AD0,missing-method}
\alias{events,ContractType,character,RiskFactorConnector-method}
\alias{events,ContractType,AD0,RiskFactorConnector-method}
\alias{events,Portfolio,character,missing-method}
\alias{events,Portfolio,AD0,missing-method}
\alias{events,Portfolio,character,RiskFactorConnector-method}
\alias{events,Portfolio,AD0,RiskFactorConnector-method}
\title{Derive the events for a \code{ContractType}}
\usage{
events(object, ad, model)

\S4method{events}{ContractType,character,missing}(object, ad, model)

\S4method{events}{ContractType,AD0,missing}(object, ad, model)

\S4method{events}{ContractType,character,RiskFactorConnector}(object, ad, model)

\S4method{events}{ContractType,AD0,RiskFactorConnector}(object, ad, model)

\S4method{events}{Portfolio,character,missing}(object, ad, model)

\S4method{events}{Portfolio,AD0,missing}(object, ad, model)

\S4method{events}{Portfolio,character,RiskFactorConnector}(object, ad, model)

\S4method{events}{Portfolio,AD0,RiskFactorConnector}(object, ad, model)
}
\arguments{
\item{object}{The \code{ContractType} for which to derive the events}

\item{ad}{The analysis date as per which all future events are to be derived}

\item{model}{(optional) The \code{RiskFactorConnector} conditional to which events are computed}
}
\value{
A \code{EventSeries} object containing the resulting events
}
\description{
The events of a contract are directly derived from the legal obligations 
defined therein. Events mark the exchange of a cash flow among the parties
involved in the contract or changes in the inner states of the contract
affecting cash flows to be exchanged in future. As such, events build the
basic analytical results upon which many financial analyses build.
}
\examples{
# define the contract
pam = Pam(
           what=list(
                 ContractID = "001",
                 Currency = "CHF",
                 ContractRole = "RPA",               
                 StatusDate       = "2016-05-30T00",
                 ContractDealDate = "2016-05-30T00",
                 InitialExchangeDate = "2016-05-30T00",
                 MaturityDate = "2020-06-01T00",
                 NotionalPrincipal = 1000,
                 NominalInterestRate = 0.05,
                 CycleOfInterestPayment = "1Y-",
                 CycleOfRateReset = "1Y-",
                 MarketObjectCodeRateReset = "RiskFreeYC",
                 DayCountConvention = "30E/360"))
                 
# define the analysis date
ad = "2016-06-01T00"

# define the risk factor connector
yc = YieldCurve(
what = list(MarketObjectCode = "RiskFreeYC",
  Nodes = list(ReferenceDate = ad, 
  Tenors = c("1W", "1M", "6M", "1Y", "2Y", "5Y"), 
  Rates = c(0.001, 0.0015, 0.002, 0.01, 0.02, 0.03))))
rf <- RFConn()
add(rf,yc)

# compute events
evs=events(pam,ad,rf)
evs
as.data.frame(evs)

}
\seealso{
\link{ContractType}, \link{EventSeries}, \link{RiskFactorConnector}
}
