% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReferenceIndex.R
\docType{methods}
\name{Index}
\alias{Index}
\alias{Index-method}
\alias{Index,numeric,ANY,character-method}
\title{\code{ReferenceIndex}-class constructor}
\usage{
Index(data, charvec, label, ...)

\S4method{Index}{numeric,ANY,character}(data, charvec, label, ...)
}
\arguments{
\item{data}{numeric vector of values for the \code{ReferenceIndex}}

\item{charvec}{Date vector}

\item{label}{ID of the \code{ReferenceIndex} class}
}
\value{
An object of class \code{ReferenceIndex}
}
\description{
Create an instance of \code{ReferenceIndex} class.
}
\seealso{
\code{\link{YieldCurve, ForeignExchangeRate}}
}
