% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DynamicYieldCurve.R
\docType{class}
\name{Interpolator-class}
\alias{Interpolator-class}
\alias{Interpolator}
\title{A Reference Class}
\description{
It should be generalized by using \code{splinefun}. 
In this case, defining a class has real value.
}
\details{
Note: A class of the same name is defined in 'YieldCurve.R' 
but only used here.
}

