% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Operations.R
\name{Investments}
\alias{Investments}
\title{Investments Contract class definition}
\usage{
Investments(ContractID, pattern, args, ...)
}
\arguments{
\item{pattern}{A function evaluating the pattern of generated investments.}

\item{args}{The list of arguments used when evaluating the pattern}
}
\description{
An Investments contract represents any operational activity 
in monetary units within an organization related to investments.
}
\details{
The \code{pattern} describes the value of the initial investment 
and the write-off pattern as \code{\link{timeSeries}} object. If the write-off
does not go down to zere the remaining value is interpreted as salvage value.
This is represented by a last event of type MD with a cash inflow of this
amount so that the \code{\link{EventSeries}} ends with a nominal value of zero.
}
\examples{
times = timeSequence(from="2014-01-01", by="3 months", length.out=9)
write.off <- function(times) {
   timeSeries(seq(1000000, 0, length.out=9), times)
   }
invest <- Investments(
              ContractID = "Invest01", Currency = "CHF", 
              pattern = write.off, 
              args = list(times = times))

}
