% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irr.R
\name{irr}
\alias{irr}
\title{\code{irr}}
\usage{
irr(object, curve, method, period, ...)
}
\arguments{
\item{object}{a timeSeries object representing cash flows.}

\item{method}{(optional) character indicating the type of discounting. Can be
one of "continuous", "compound" or "linear".}

\item{period}{(optional) character indicating the day count convention.}
}
\value{
the internal rate of return calculated from the cash flows.
}
\description{
Function computing the internal rate of return of a series of cash flows.
Computation of the internal rate of return requires the numerical
determination of the zeroes of a polynominal. In general these may be complex. 
A unique real solution is ensured if the first cash flow is negative and
all the following cash flows are non-negative with at least one of them 
being strictly positive, or the first cash flow is positive and all of the
followings are non-positive with at least one of them being strictly negative.
}
\examples{
# timeSeries object
cfs <- timeSeries(data = c(-100, 5, 5, 105), 
                  charvec = c("2014-01-01","2015-01-01","2016-01-01","2017-01-01"))
r <- irr(cfs)

# Bond
b <- bond("2013-12-31", maturity = "5 years", nominal = 50000, 
           coupon = 0.02, couponFreq = "1 years")
irr(b)

# Annuity
ann <- annuity("2013-12-31", nominal = 50000, ir = 0.02, maturity = "5 years")
irr(ann)
          
}
