% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ContractPlot.R, R/RiskFactorPlots.R
\docType{methods}
\name{plot,ContractType,character-method}
\alias{plot,ContractType,character-method}
\alias{plot,FEMSContract,character-method}
\alias{plot,YieldCurve,missing-method}
\alias{plot,RiskFactor,ANY-method}
\title{Plot a \link{YieldCurve}}
\usage{
\S4method{plot}{ContractType,character}(x, y, yc = NULL, ...)

\S4method{plot}{FEMSContract,character}(x, y, yc = NULL, to = NULL, ...)

\S4method{plot}{YieldCurve,missing}(x, y, ...)

\S4method{plot}{RiskFactor,ANY}(x, y, ...)
}
\arguments{
\item{object}{The \link{RiskFactor} object to plot}
}
\value{



}
\description{
Create a graphical representation of a \link{YieldCurve}-
object.

Create a graphical representation of an index-like 
\link{RiskFactor}-object.
}
\examples{
yc <- YieldCurve()
tenors <- c("1W", "1M", "6M", "1Y", "2Y", "5Y")
rates <- c(0.001, 0.0015, 0.002, 0.01, 0.02, 0.03)
set(yc, what = list(MarketObjectCode = "YC_Prim",
  Nodes = list(ReferenceDate = "2015-01-01T00", Tenors = tenors, Rates = rates)))
plot(yc)

idx <- Index()
times <- c("2015-01-01T00", "2016-01-01T00", "2017-01-01T00", "2018-01-01T00",
           "2019-01-01T00")
values <- c(100, 110, 120, 130, 140)
set(idx, what=list(
  MarketObjectCode = "IND_CPI_EA",
  Data=list(Dates=times,Values=values)))
plot(idx)

}
\seealso{
\link{YieldCurve},\link{plot}

\link{ReferenceIndex},\link{plot}
}
