% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YieldCurve.R
\docType{methods}
\name{YieldCurve}
\alias{YieldCurve}
\alias{YieldCurve-method}
\alias{YieldCurve,ANY-method}
\title{\code{YieldCurve}-class constructor}
\usage{
YieldCurve(...)

\S4method{YieldCurve}{ANY}(...)
}
\arguments{
\item{...}{}
}
\value{
An object of class \code{YieldCurve} 
         containing the reference to the Java object
}
\description{
Create an instance of \code{YieldCurve} class. The 
constructor will also create an instance of the respective
Java class in the running JVM.
}
\seealso{
\code{\link{ReferenceIndex, ForeignExchangeRate}}
}
